==> archive2/analysis_merfish/35.integrated_cell_level-Copy5.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import os

import anndata as ad
import scanpy as sc

from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from umap import UMAP


from scroutines import basicu
from scroutines import powerplots
from scroutines.miscu import is_in_polygon

import utils_merfish
from merfish_datasets import merfish_datasets
from merfish_genesets import get_all_genesets

import importlib
importlib.reload(powerplots)
importlib.reload(utils_merfish)

np.random.seed(0)

def binning_pipe(adata, n=20, layer='lnorm', bin_type='depth_bin'):
    """
    """
    assert bin_type in ['depth_bin', 'width_bin']
    # bin it 
    depth_bins, depth_binned = utils_merfish.binning(adata.obs['depth'].values, n)
    width_bins, width_binned = utils_merfish.binning(adata.obs['width'].values, n)

    norm_ = pd.DataFrame(adata.layers[layer], columns=adata.var.index)
    norm_['depth_bin'] = depth_binned
    norm_['width_bin'] = width_binned
    
    norm_mean = norm_.groupby(bin_type).mean(numeric_only=True)
    norm_sem  = norm_.groupby(bin_type).sem(numeric_only=True)
    norm_std  = norm_.groupby(bin_type).std(numeric_only=True)
    norm_n    = norm_[bin_type].value_counts(sort=False)

    return norm_mean, norm_sem, norm_std, norm_n, depth_binned, width_binned, depth_bins, width_bins

genesets = get_all_genesets()
genesets

directories = merfish_datasets
print(merfish_datasets)

ddir = "/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish/organized" 
fout = os.path.join(ddir, 'P28NRDR_v1l23_merged_240508.h5ad')
!ls $ddir/*240411.h5ad 

%%time
names = [
    'P28NR_ant', 
    'P28NR_pos',
    
    'P28NRb_ant', 
    'P28NRb_pos',
    
    'P28DR_ant', 
    'P28DR_pos',
    
    'P28DRb_ant', 
    'P28DRb_pos',
]

alldata = {}
for name in names:
    if 'P28NRb_ant' in name:
        alldata[name] = ad.read(os.path.join(ddir, f'{name}_ctxglut_240506.h5ad'))
    elif 'b' in name:
        alldata[name] = ad.read(os.path.join(ddir, f'{name}_ctxglut_240429.h5ad'))
    else:
        alldata[name] = ad.read(os.path.join(ddir, f'{name}_ctxglut_240411.h5ad'))
    print(name, len(alldata[name]))
    


for name, adata in alldata.items():
    print(name)
    gns = ['Scnn1a', 'Rorb', 'Igfbp4', 'Fos', 'Sorcs3']
    x = adata.obs['x']
    y = adata.obs['y']
    n = len(gns)

    fig, axs = plt.subplots(1,n,figsize=(n*6,1*5))
    for ax, gn in zip(axs, gns):
        g = adata[:,gn].layers['norm']
        utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=1, title=gn)
    plt.show()

# viz

v1l23_data = {}
for name in names:
    adata = alldata[name]
    # bound_l, bound_r, bound_d = adata.uns['bound_lrd']
    adatasub = adata[adata.obs['inside_v1l23']]
    v1l23_data[name] = adatasub

gns = ['Scnn1a', 'Rorb', 'Igfbp4', 'Whrn', 'Fos', 'Cdh13', 'Sorcs3', 'Chrm2',]# 'Astn2', 'Fos']

name = 'P28DRb_pos' 
adata = alldata[name]
bound_l, bound_r, bound_d = adata.uns['bound_lrd']
x =  adata.obs['width']
y = -adata.obs['depth']
n = len(gns)

fig, axs = plt.subplots(n,1,figsize=(1*10,n*1.5))
fig.suptitle(name)
for ax, gn in zip(axs, gns):
    # gn = 'Scnn1a'
    g = np.log2(1+adata[:,gn].layers['norm']).reshape(-1,)
    
    vmax = np.percentile(g, 99)
    vmin = np.percentile(g,  5)
    cond = g > 1e-5
    
    utils_merfish.st_scatter_ax(fig, ax, x[~cond], y[~cond], gexp=g[~cond], s=2, title='', vmin=vmin, vmax=vmax, cmap='rocket_r') #, axis_off=False)
    utils_merfish.st_scatter_ax(fig, ax,  x[cond],  y[cond],  gexp=g[cond], s=2, title='', vmin=vmin, vmax=vmax, cmap='rocket_r') #, axis_off=False)
    
    ax.set_title(gn, loc='left', ha='right', y=0.5)
    ax.axhline(-bound_d, linestyle='--', linewidth=1, color='gray', zorder=2)
    ax.axvline( bound_l, linestyle='--', linewidth=1, color='gray', zorder=2)
    ax.axvline( bound_r, linestyle='--', linewidth=1, color='gray', zorder=2)
    # ax.grid(True)
    # ax.axis('on')
    
plt.show()

gns = ['Fos', 'Cdh13', 'Sorcs3', 'Chrm2']
densities = []
for name in names:
    adata = v1l23_data[name]
    x =  adata.obs['width']
    y = -adata.obs['depth']
    n = len(gns)
    width_range = adata.obs['width'].max() - adata.obs['width'].min()
    print(name, len(adata), width_range, len(adata)/width_range)
    densities.append(len(adata)/width_range)

    fig, axs = plt.subplots(1,n,figsize=(n*6,1*1))
    for ax, gn in zip(axs, gns):
        # gn = 'Scnn1a'
        g = np.log2(1+adata[:,gn].layers['norm'])
        utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=5, title=gn)
        ax.hlines(-np.array([0,100,200,300,400]), np.min(x)-100, np.min(x)-10, color='k', linestyle='--')
        ax.set_aspect('equal')

    plt.show()

# do a clustering together
- abc genes

agenes = genesets['a']
bgenes = genesets['b']
cgenes = genesets['c']
iegs   = genesets['i']
up_agenes = genesets['a_up']
abcgenes = np.hstack([agenes, bgenes, cgenes])
len(abcgenes)

agenes_idx = basicu.get_index_from_array(adatasub.var.index.values, agenes)
bgenes_idx = basicu.get_index_from_array(adatasub.var.index.values, bgenes)
cgenes_idx = basicu.get_index_from_array(adatasub.var.index.values, cgenes)
igenes_idx = basicu.get_index_from_array(adatasub.var.index.values, iegs)

mean_total_rna_target = 250
adata_premerge = []
for i, name in enumerate(names):
    adatasub = v1l23_data[name].copy()
    
    adatasub.obs['sample'] = name
    
    norm_cnts = adatasub.layers['norm']
    mean_per_batch = np.mean(norm_cnts.sum(axis=1))
    adatasub.layers['jnorm']  = norm_cnts*(mean_total_rna_target/mean_per_batch)
    adatasub.layers['ljnorm'] = np.log2(1+adatasub.layers['jnorm'])
    
    adatasub.obs['norm_transcript_count']  = adatasub.layers['norm'].sum(axis=1)
    adatasub.obs['jnorm_transcript_count'] = adatasub.layers['jnorm'].sum(axis=1)
    
    adata_premerge.append(adatasub)
adata_premerge = ad.concat(adata_premerge)

xcol, ycol = 'volume', 'transcript_count'
fig, axs = plt.subplots(2,4,figsize=(4*5,2*4), sharex=True, sharey=True)
for i, name in enumerate(names):
    ax = axs.flat[i]
    
    vol = adata_premerge.obs[xcol]
    cnt = adata_premerge.obs[ycol]
    ax.scatter(np.log10(vol), np.log10(cnt), s=1, color='lightgray')
    
    adatasub = adata_premerge[adata_premerge.obs['sample']==name]
    vol = adatasub.obs[xcol]
    cnt = adatasub.obs[ycol]
    ax.scatter(np.log10(vol), np.log10(cnt), s=1)
    ax.set_title(name)
    
ax = axs.flat[0]
ax.set_xlabel('log10(volume)')
ax.set_ylabel('log10(trscpt count)')
plt.show()
    
#     sns.histplot(vol, bins=np.linspace(0,1000,50), 
    

xcol, ycol = 'volume', 'jnorm_transcript_count'
ccol = 'transcript_count'
fig, axs = plt.subplots(2,4,figsize=(4*5,2*4), sharex=True, sharey=True)
for i, name in enumerate(names):
    ax = axs.flat[i]
    
    vol = adata_premerge.obs[xcol]
    cnt = adata_premerge.obs[ycol]
    ax.scatter(np.log10(vol), np.log10(cnt), s=1, color='lightgray')
    
    adatasub = adata_premerge[adata_premerge.obs['sample']==name]
    vol = adatasub.obs[xcol]
    cnt = adatasub.obs[ycol]
    ocnt = adatasub.obs[ccol]
    ax.set_title(name)
    
    p = ax.scatter(np.log10(vol), np.log10(cnt), c=np.log10(ocnt), s=1)
    fig.colorbar(p)
    
ax = axs.flat[0]
ax.set_xlabel('log10(volume)')
ax.set_ylabel('log10(trscpt count)')
plt.show()
    
#     sns.histplot(vol, bins=np.linspace(0,1000,50), 
    

xcol, ycol = 'volume', 'jnorm_transcript_count'
ccol = 'transcript_count'
fig, axs = plt.subplots(2,4,figsize=(4*5,2*4), sharex=True, sharey=True)
for i, name in enumerate(names):
    ax = axs.flat[i]
    
    vol = adata_premerge.obs[xcol]
    cnt = adata_premerge.obs[ycol]
    ax.scatter(vol, cnt, s=1, color='lightgray')
    
    adatasub = adata_premerge[adata_premerge.obs['sample']==name]
    vol = adatasub.obs[xcol]
    cnt = adatasub.obs[ycol]
    ocnt = adatasub.obs[ccol]
    ax.set_title(name)
    
    p = ax.scatter(vol, cnt, s=1)
    # fig.colorbar(p)
    ax.set_xlim([0,500])
    
ax = axs.flat[0]
ax.set_xlabel('log10(volume)')
ax.set_ylabel('log10(trscpt count)')
plt.show()
    
#     sns.histplot(vol, bins=np.linspace(0,1000,50), 
    

xcol, ycol = 'jnorm_transcript_count', 'transcript_count'
fig, axs = plt.subplots(2,4,figsize=(4*5,2*4), sharex=True, sharey=True)
for i, name in enumerate(names):
    ax = axs.flat[i]
    
    vol = adata_premerge.obs[xcol]
    cnt = adata_premerge.obs[ycol]
    ax.scatter(np.log10(vol), np.log10(cnt), s=1, color='lightgray')
    
    adatasub = adata_premerge[adata_premerge.obs['sample']==name]
    vol = adatasub.obs[xcol]
    cnt = adatasub.obs[ycol]
    ax.set_title(name)
    
    ax.scatter(np.log10(vol), np.log10(cnt), s=1)
    
ax = axs.flat[0]
ax.set_xlabel('log10(volume)')
ax.set_ylabel('log10(trscpt count)')
plt.show()
    
#     sns.histplot(vol, bins=np.linspace(0,1000,50), 
    

sns.boxplot(data=adata_premerge.obs, x='sample', y='transcript_count')

sns.boxplot(data=adata_premerge.obs, x='sample', y='norm_transcript_count')

sns.boxplot(data=adata_premerge.obs, x='sample', y='jnorm_transcript_count')

sns.histplot(adata_premerge.obs['jnorm_transcript_count'], bins=np.linspace(0, 500, 50))

# filter
# adatasub = adatasub[adatasub.obs['transcript_count'] > 30]
metric0 = adata_premerge.obs['volume'].values
metric1 = adata_premerge.obs['transcript_count'].values
metric2 = adata_premerge.obs['jnorm_transcript_count'].values
print(np.percentile(metric1, 10), 
      np.percentile(metric1, 20),
      np.percentile(metric1, 30),
     )
print(np.percentile(metric2, 10), 
      np.percentile(metric2, 20),
      np.percentile(metric2, 30),
     )
# cond = metric2 > 100
# cond = metric1 > 50
# cond = np.logical_or(metric1 > 50, metric2 > 100)
# cond = np.logical_and(metric1 > 50, metric2 > 100)
# cond = np.logical_and(metric1 > 30, metric2 > 70)
# cond = metric1 > 50
cond = metric0 > 100

# cutmin, cutmax = np.percentile(metric, 10), np.percentile(metric, 99)
# adata_premerge2 = adata_premerge[np.logical_and(metric > cutmin, metric < cutmax)]
adata_premerge2 = adata_premerge[cond].copy()

# cluster V1 L2/3 only
n0 = len(adata_premerge)
n1 = len(adata_premerge2)
print(f'{n1/n0*100:.1f}')


metric0 = adata_premerge2.obs['volume'].values
metric1 = adata_premerge2.obs['transcript_count'].values
metric2 = adata_premerge2.obs['jnorm_transcript_count'].values
print(np.percentile(metric1, 10), 
      np.percentile(metric1, 20),
      np.percentile(metric1, 30),
     )
print(np.percentile(metric2, 10), 
      np.percentile(metric2, 20),
      np.percentile(metric2, 30),
     )
# cond = metric1 > np.percentile(metric1, 10)
cond = np.logical_or(
    metric1 > np.percentile(metric1, 20), 
    metric2 > np.percentile(metric2, 20),
    )
# cond = np.logical_and(
#     metric1 > np.percentile(metric1, 10), 
#     metric2 > np.percentile(metric2, 10),
#     )
adata_premerge2 = adata_premerge2[cond].copy()
n1 = len(adata_premerge2)
print(f'{n1/n0*100:.1f}')



adata_merged = []
for i, name in enumerate(names):
    adatasub = adata_premerge2[adata_premerge2.obs['sample']==name].copy()
        
    adatasub.obs.index = adatasub.obs.index + '_' + name
    adatasub.obs['sample'] = name
    adatasub.obs['depth_show'] = -adatasub.obs['depth'].values - i*500 # name
    adatasub.obs['width_show'] =  adatasub.obs['width'].values - np.min(adatasub.obs['width'].values)   # name
    adata_merged.append(adatasub)
    print(adatasub.shape)
    # break
    
adata_merged = ad.concat(adata_merged)
adata_merged_abcgenes = adata_merged[:,abcgenes].copy()

adata_merged, adata_merged_abcgenes

# adata = adata_merged
# adata = adata_premerge.copy()
adata = adata_premerge2.copy()

sns.histplot(adata.obs['jnorm_transcript_count'].values, bins=np.linspace(0,200,100))

sns.histplot(adata.obs['transcript_count'].values, bins=np.linspace(0,200,100))

fig, ax = plt.subplots()
sns.boxplot(data=adata.obs, y='jnorm_transcript_count', x='sample', ax=ax) #].values
ax.set_xticklabels(ax.get_xticklabels(), rotation=90)

dfwork = adata.obs.copy() 
xlbl = 'volume'
ylbl = 'transcript_count'

x = dfwork[xlbl].values 
y = dfwork[ylbl].values
bins = np.linspace(0,1000,50)
dfwork['bin'] = pd.cut(x, bins)
res = dfwork.groupby('bin')[[xlbl, ylbl]].mean()


plt.scatter(x,y, s=1, edgecolor='none')
# sns.kdeplot(data=dfwork, x=xlbl, y=ylbl, bins=bins)
plt.plot(res[xlbl], res[ylbl], color='k')
plt.xlabel(xlbl)
plt.ylabel(ylbl)
# plt.xlim([0,200])
# plt.ylim([0,500])

dfwork = adata.obs.copy() 
xlbl = 'volume'
ylbl = 'jnorm_transcript_count'

x = dfwork[xlbl].values 
y = dfwork[ylbl].values
bins = np.linspace(0,500,50)
dfwork['bin'] = pd.cut(x, bins)
res = dfwork.groupby('bin')[[xlbl, ylbl]].mean()


plt.scatter(x,y, s=1, edgecolor='none')
# sns.kdeplot(data=dfwork, x=xlbl, y=ylbl, bins=bins)
plt.plot(res[xlbl], res[ylbl], color='k')
plt.xlabel(xlbl)
plt.ylabel(ylbl)
# plt.xlim([0,400])
plt.xlim([0,500])

dfwork = adata.obs.copy() 
xlbl = 'transcript_count'
ylbl = 'jnorm_transcript_count'
clbl = 'volume'

x = dfwork[xlbl].values 
y = dfwork[ylbl].values
bins = np.linspace(0,500,50)
dfwork['bin'] = pd.cut(x, bins)
res = dfwork.groupby('bin')[[xlbl, ylbl]].mean()


p = plt.scatter(x,y, c=dfwork[clbl], s=1, edgecolor='none', vmax=500, cmap='viridis')
plt.colorbar(p)
# sns.kdeplot(data=dfwork, x=xlbl, y=ylbl, bins=bins)
plt.plot(res[xlbl], res[ylbl], color='k')
plt.xlabel(xlbl)
plt.ylabel(ylbl)
plt.xlim([0,500])
# plt.ylim([0,500])

dfwork = adata.obs.copy() 
xlbl = 'transcript_count'
ylbl = 'jnorm_transcript_count'
clbl = 'volume'

x = dfwork[xlbl].values 
y = dfwork[ylbl].values
c = dfwork[clbl].values
bins = np.linspace(0,500,50)
dfwork['bin'] = pd.cut(x, bins)
res = dfwork.groupby('bin')[[xlbl, ylbl]].mean()


p = plt.scatter(x,y, c=c, s=1, edgecolor='none', vmax=100, cmap='viridis')
plt.colorbar(p)
# sns.kdeplot(data=dfwork, x=xlbl, y=ylbl, bins=bins)
plt.plot(res[xlbl], res[ylbl], color='k')
plt.xlabel(xlbl)
plt.ylabel(ylbl)
plt.xlim([0,200])
plt.ylim([0,500])

dfwork = adata.obs.copy() 
xlbl = 'transcript_count'
ylbl = 'jnorm_transcript_count'
clbl = 'volume'

x = dfwork[xlbl].values 
y = dfwork[ylbl].values
c = dfwork[clbl].values
bins = np.linspace(0,1000,50)
dfwork['bin'] = pd.cut(x, bins)
res = dfwork.groupby('bin')[[xlbl, ylbl]].mean()


plt.scatter(x, y, c=c, s=1, edgecolor='none', cmap='viridis', vmax=110)
plt.plot(res[xlbl], res[ylbl], color='k')
plt.xlabel(xlbl)
plt.ylabel(ylbl)
# plt.xlim([0,1000])
# plt.ylim([0,500])


from scipy.stats import zscore

marker_genes = [
       'Ptprn', 'Slc17a7', 'Gad1', 'Fos', 
       
       'Gfap', 'Slc6a13', 'Slc47a1',
       'Grin2c', 'Aqp4', 'Rfx4', 'Sox21', 'Slc1a3',
       
       'Sox10', 'Pdgfra', 'Mog',
       
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
      ] 

genes_noniegs = np.array([g for g in adata.var.index.values if g not in iegs])
genes_noniegs.shape

adata = adata_merged # [:,marker_genes].copy()


# PCA
pca = PCA(n_components=10)
pcs = pca.fit_transform(zscore(adata[:,genes_noniegs].layers['ljnorm'], axis=1))
# pcs = pca.fit_transform(zscore(adata.layers['ljnorm'], axis=1))
ucs = UMAP(n_components=2, n_neighbors=30, random_state=0).fit_transform(pcs)

adata.obsm['pca'] = pcs
adata.obsm['umap'] = ucs
sc.pp.neighbors(adata, n_neighbors=30, use_rep='pca', random_state=0)

# clustering
r = 0.3
sc.tl.leiden(adata, resolution=r, key_added=f'leiden_r{r}', random_state=0, n_iterations=10)

# plot
gn = 'Slc17a7'
# gn = 'Fos'
# gn = 'Gad1'
g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))

# add some quality metrics

fig, (ax1, ax2) = plt.subplots(1,2,figsize=(2*5,1*4))
utils_merfish.st_scatter_ax(fig, ax1, pcs[:,0], pcs[:,1], gexp=g)
utils_merfish.st_scatter_ax(fig, ax2, ucs[:,0], ucs[:,1], gexp=g)
plt.show()

# plot
gn = 'Slc17a7'
# gn = 'Fos'
# gn = 'Gad1'
g = adata[:,gn].layers['jnorm'].reshape(-1,)

# add some quality metrics

fig, (ax1, ax2) = plt.subplots(1,2,figsize=(2*5,1*4))
utils_merfish.st_scatter_ax(fig, ax1, pcs[:,0], pcs[:,1], gexp=g)
utils_merfish.st_scatter_ax(fig, ax2, ucs[:,0], ucs[:,1], gexp=g)
plt.show()


clsts = adata.obs[f'leiden_r{r}'].astype(int)
xr =  adata.obs['width_show']
yr =  adata.obs['depth_show']
ux    = adata.obsm['umap'][:,0]
uy    = adata.obsm['umap'][:,1]
utils_merfish.plot_cluster(clsts, xr, yr, ux, uy, s=2)

samples, uniq_labels = pd.factorize(adata.obs['sample']) # .astype(int)
utils_merfish.plot_cluster(samples, xr, yr, ux, uy, s=2)

np.unique(clsts, return_counts=True)

clsts = adata.obs[f'leiden_r{r}'].astype(int)
uniq_clsts = np.unique(clsts)


for clst in uniq_clsts:
    show = (clsts == clst)
    xr =  adata.obs['width_show']
    yr =  adata.obs['depth_show']
    ux    = adata.obsm['umap'][:,0]
    uy    = adata.obsm['umap'][:,1]
    utils_merfish.plot_cluster(show, xr, yr, ux, uy, s=2, cmap=plt.cm.copper_r, suptitle=clst)

# figure out major cell population by marker genes; groups of genes; and quality metrics

# plot
marker_genes = [
       'Ptprn', 'Slc17a7', 'Gad1', 'Fos', 
       
       'Gfap', 'Slc6a13', 'Slc47a1',
       'Grin2c', 'Aqp4', 'Rfx4', 'Sox21', 'Slc1a3',
       
       'Sox10', 'Pdgfra', 'Mog',
       
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
      ] 
gns = marker_genes
n = len(gns)
nx = 4
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for gn, ax in zip(gns, axs.flat):
    g = adata[:,gn].layers['jnorm'].reshape(-1,)
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g)
    ax.set_title(gn)
plt.show()


# plot
marker_genes = [
       'Ptprn', 'Slc17a7', 'Gad1', 'Fos', 
       
       'Gfap', 'Slc6a13', 'Slc47a1',
       'Grin2c', 'Aqp4', 'Rfx4', 'Sox21', 'Slc1a3',
       
       'Sox10', 'Pdgfra', 'Mog',
       
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
      ] 
gns = marker_genes
n = len(gns)
nx = 4
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for gn, ax in zip(gns, axs.flat):
    g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g)
    ax.set_title(gn)
plt.show()



# # add some quality metrics
# fig, ax = plt.subplots()
# # g = (adata.layers['jnorm'].sum(axis=1) < 100).astype(int)
# g = (adata.obs['jnorm_transcript_count'].values < 100).astype(int)
# # g = (adata.obs['volume'].values < 60).astype(int)
# p = utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
# fig.colorbar(p)
# ax.set_title('')
# plt.show()

# fig, ax = plt.subplots()
# # g = (adata.layers['volume'].sum(axis=1)) #  < 100).astype(int)
# # g = (adata.obs['transcript_count'].values < 50).astype(int)
# g = (adata.obs['volume'].values < 150).astype(int)
# p = utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
# fig.colorbar(p)
# ax.set_title('')
# plt.show()

# fig, ax = plt.subplots()
# # g = (adata.layers['volume'].sum(axis=1)) #  < 100).astype(int)
# g = (adata.obs['transcript_count'].values < 100).astype(int)
# # g = (adata.obs['volume'].values < 150).astype(int)
# p = utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
# fig.colorbar(p)
# ax.set_title('')
# plt.show()

fig, ax = plt.subplots()
# g = (adata.layers['volume'].sum(axis=1)) #  < 100).astype(int)
g = np.logical_and(
    (adata.obs['transcript_count'].values < 50).astype(int),
    (adata.obs['jnorm_transcript_count'].values < 150).astype(int),
).astype(int)
# g = (adata.obs['volume'].values < 150).astype(int)
p = utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
fig.colorbar(p)
ax.set_title('')
plt.show()

metrics = [
    'volume', 'anisotropy', 'perimeter_area_ratio', 'solidity', 
    'PolyT_raw', 'PolyT_high_pass', 'DAPI_raw', 'DAPI_high_pass', 
    'transcript_count', 'jnorm_transcript_count', 'gnnum', 'fpcov', 
    'depth', 'width', 'sample' 
       ]
n = len(metrics)
nx = 5
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for metric, ax in zip(metrics, axs.flat):
    g = adata.obs[metric].values
    if metric == 'sample':
        g, uniq_lbls = pd.factorize(g)
    p = utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
    fig.colorbar(p, shrink=0.4)
    ax.set_title(metric)
plt.show()


n = len(metrics)
nx = 5
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for metric, ax in zip(metrics, axs.flat):
    if metric == 'sample':
        g, uniq_lbls = pd.factorize(g)
    else:
        g = np.log10(1+adata.obs[metric].values)
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
    ax.set_title(metric)
plt.show()

adata

# broad annotation and save

np.unique(clsts, return_counts=True)

clst_annots = [
    "Glu0",
    "Glu1",
    "Glu2",
    # "Glu2",
    "3",
    "4",
    "5",
    "6",
    "7",
    "8",
    # "9",
    # "10",
    # 'Glu',
    # 'Glu',
    # 'low',
    # 'Glu',
    # 'micro', 
    # 'GABA',
    # 'astro',
    # 'VLMC',
    # 'Olig',
    
     # 'Glu_nr_low', 
     # 'olig_and_low_qual',
     # 'Glu_dr_low', 
     # 'Glu_nr_high', 
     # 'Glu_dr_high', 
     # 'GABA',
]
adata.uns['clst_annots'] = clst_annots

print(fout)

adata.write(fout)

adata.obs
