==> archive2/analysis_merfish/42.v1l23_merged_analysis-plots.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import os

import anndata as ad
import scanpy as sc

from scipy import stats
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from umap import UMAP


from scroutines import basicu
from scroutines import powerplots
from scroutines.miscu import is_in_polygon

import utils_merfish
from merfish_datasets import merfish_datasets
from merfish_genesets import get_all_genesets

import importlib
importlib.reload(powerplots)
importlib.reload(utils_merfish)
import merfish_genesets
importlib.reload(merfish_genesets)
from merfish_genesets import get_all_genesets

np.random.seed(0)

def binning_pipe(adata, n=20, layer='lnorm', bin_type='depth_bin'):
    """
    """
    assert bin_type in ['depth_bin', 'width_bin']
    # bin it 
    depth_bins, depth_binned = utils_merfish.binning(adata.obs['depth'].values, n)
    width_bins, width_binned = utils_merfish.binning(adata.obs['width'].values, n)

    norm_ = pd.DataFrame(adata.layers[layer], columns=adata.var.index)
    norm_['depth_bin'] = depth_binned
    norm_['width_bin'] = width_binned
    
    norm_mean = norm_.groupby(bin_type).mean(numeric_only=True)
    norm_sem  = norm_.groupby(bin_type).sem(numeric_only=True)
    norm_std  = norm_.groupby(bin_type).std(numeric_only=True)
    norm_n    = norm_[bin_type].value_counts(sort=False)

    return norm_mean, norm_sem, norm_std, norm_n, depth_binned, width_binned, depth_bins, width_bins

def binning_pipe2(adata, col_to_bin, layer, bins=None, n=20):
    """
    """
    if bins is None:
        # bin it 
        bins, binned = utils_merfish.binning(adata.obs[col_to_bin].values, n)
    else:
        binned = pd.cut(adata.obs[col_to_bin].values, bins=bins)

    norm_ = pd.DataFrame(adata.layers[layer], columns=adata.var.index)
    norm_['thebin'] = binned
    
    norm_mean = norm_.groupby('thebin').mean(numeric_only=True)
    norm_sem  = norm_.groupby('thebin').sem(numeric_only=True)
    norm_std  = norm_.groupby('thebin').std(numeric_only=True)
    norm_n    = norm_['thebin'].value_counts(sort=False)

    return norm_mean, norm_sem, norm_std, norm_n, binned, bins 

genesets, typegenes_df = get_all_genesets()
genesets
for key, item in genesets.items():
    print(key, len(item))

typegenes_df

agenes = genesets['a']
bgenes = genesets['b']
cgenes = genesets['c']
iegs   = genesets['i']
up_agenes = genesets['a_up']
abcgenes = np.hstack([agenes, bgenes, cgenes])
len(abcgenes), len(iegs)

directories = merfish_datasets
print(merfish_datasets)

ddir = "/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish/organized" 
!ls $ddir/*240411.h5ad 

%%time
fin = os.path.join(ddir, 'P28NRDR_v1l23_merged_240508.h5ad')
adata = ad.read(fin) 

clst_annots = adata.uns['clst_annots']
pcs = adata.obsm['pca']
ucs = adata.obsm['umap']
r = 0.3

names = [
    'P28NR_ant', 
    'P28NR_pos',
    'P28NRb_ant', 
    'P28NRb_pos',
    
    'P28DR_ant', 
    'P28DR_pos',
    'P28DRb_ant', 
    'P28DRb_pos',
]
adata

agenes_idx = basicu.get_index_from_array(adata.var.index.values, agenes)
bgenes_idx = basicu.get_index_from_array(adata.var.index.values, bgenes)
cgenes_idx = basicu.get_index_from_array(adata.var.index.values, cgenes)
igenes_idx = basicu.get_index_from_array(adata.var.index.values, iegs)

# plot everything 


clsts = adata.obs[f'leiden_r{r}'].astype(int)
xr =  adata.obs['width_show']
yr =  adata.obs['depth_show']
ux    = adata.obsm['umap'][:,0]
uy    = adata.obsm['umap'][:,1]
utils_merfish.plot_cluster(clsts, xr, yr, ux, uy, s=2)

samples, uniq_labels = pd.factorize(adata.obs['sample']) # .astype(int)
utils_merfish.plot_cluster(samples, xr, yr, ux, uy, s=2)

np.unique(clsts, return_counts=True)

clsts = adata.obs[f'leiden_r{r}'].astype(int)
uniq_clsts = np.unique(clsts)
n = len(uniq_clsts)

fig, axs = plt.subplots(1,n,figsize=(n*4,1*4))
for i, clst in enumerate(uniq_clsts):
    show = (clsts == clst)
    xr =  adata.obs['width_show']
    yr =  adata.obs['depth_show']
    ux    = adata.obsm['umap'][:,0]
    uy    = adata.obsm['umap'][:,1]
    
    ax = axs[i]
    ax.scatter(ux, uy, c=show, s=1, edgecolor='none', cmap='rocket_r', rasterized=True)
    ax.set_aspect('equal')
    ax.axis('off')
    ax.set_title(f"C{clst}")
    
fig, axs = plt.subplots(1,n,figsize=(n*4,1*6))
for i, clst in enumerate(uniq_clsts):
    show = (clsts == clst)
    xr =  adata.obs['width_show']
    yr =  adata.obs['depth_show']
    ux    = adata.obsm['umap'][:,0]
    uy    = adata.obsm['umap'][:,1]
    
    ax = axs[i]
    ax.scatter(xr, yr, c=show, s=3, edgecolor='none', cmap='rocket_r', rasterized=True)
    ax.set_aspect('equal')
    ax.axis('off')
    ax.set_title(f"C{clst}")
    
    
    # utils_merfish.plot_cluster(show, xr, yr, ux, uy, s=2, cmap=plt.cm.copper_r, suptitle=clst)

clsts = adata.obs[f'leiden_r{r}'].astype(int)
uniq_clsts = np.unique(clsts)
n = len(uniq_clsts)

fig, ax = plt.subplots(1,1,figsize=(1*4,1*4))
show = [_c in [0,1,2] for _c in clsts]
xr =  adata.obs['width_show']
yr =  adata.obs['depth_show']
ux    = adata.obsm['umap'][:,0]
uy    = adata.obsm['umap'][:,1]

ax.scatter(ux, uy, c=show, s=1, edgecolor='none', cmap='rocket_r', rasterized=True)
ax.set_aspect('equal')
ax.axis('off')
ax.set_title(f"C0,C1,C2")
    
fig, ax = plt.subplots(1,1,figsize=(1*4,1*6))
show = [_c in [0,1,2] for _c in clsts]
xr =  adata.obs['width_show']
yr =  adata.obs['depth_show']
ux    = adata.obsm['umap'][:,0]
uy    = adata.obsm['umap'][:,1]

ax.scatter(xr, yr, c=show, s=3, edgecolor='none', cmap='rocket_r', rasterized=True)
ax.set_aspect('equal')
ax.axis('off')
ax.set_title(f"C0,C1,C2")


# utils_merfish.plot_cluster(show, xr, yr, ux, uy, s=2, cmap=plt.cm.copper_r, suptitle=clst)

# figure out major cell population by marker genes; groups of genes; and quality metrics

# plot
marker_genes = [
       'Ptprn', 'Slc17a7', 'Gad1', 'Fos', 
       
       'Gfap', 'Slc6a13', 'Slc47a1',
       'Grin2c', 'Aqp4', 'Rfx4', 'Sox21', 'Slc1a3',
       
       'Sox10', 'Pdgfra', 'Mog',
       
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
      ] 
gns = marker_genes
n = len(gns)
nx = 4
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for gn, ax in zip(gns, axs.flat):
    g = adata[:,gn].layers['jnorm'].reshape(-1,)
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g)
    ax.set_title(gn)
plt.show()


# plot
marker_genes = [
       'Ptprn', 'Slc17a7', 'Gad1', 'Fos', 
       
       'Gfap', 'Slc6a13', 'Slc47a1',
       'Grin2c', 'Aqp4', 'Rfx4', 'Sox21', 'Slc1a3',
       
       'Sox10', 'Pdgfra', 'Mog',
       
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
      ] 
gns = marker_genes
n = len(gns)
nx = 4
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for gn, ax in zip(gns, axs.flat):
    g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g)
    ax.set_title(gn)
plt.show()



metrics = [
    'volume', 'anisotropy', 'perimeter_area_ratio', 'solidity', 
    'PolyT_raw', 'PolyT_high_pass', 'DAPI_raw', 'DAPI_high_pass', 
    'transcript_count', 'jnorm_transcript_count', 'gnnum', 'fpcov', 
    'depth', 'width', 'sample' 
       ]
n = len(metrics)
nx = 5
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for metric, ax in zip(metrics, axs.flat):
    g = adata.obs[metric].values
    if metric == 'sample':
        g, uniq_lbls = pd.factorize(g)
    p = utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
    fig.colorbar(p, shrink=0.4)
    ax.set_title(metric)
plt.show()


metrics = [
    'volume', 'PolyT_raw', 
    'gnnum', 'transcript_count', 'jnorm_transcript_count',  # 'fpcov', 
       ]
n = len(metrics)
nx = 5
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for metric, ax in zip(metrics, axs.flat):
    g = adata.obs[metric].values
    if metric == 'sample':
        g, uniq_lbls = pd.factorize(g)
    p = utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g, s=3)
    fig.colorbar(p, shrink=0.4)
    ax.set_title(metric)
plt.show()


dfwork = adata.obs.copy() 
xlbl = 'volume'
x = dfwork[xlbl].values 
bins = np.logspace(2,3.2,50)
dfwork['bin'] = pd.cut(x, bins)

    
fig, axs = plt.subplots(1,2,figsize=(8,4))
for ax, ylbl in zip(axs, ['transcript_count', 'jnorm_transcript_count']):
    
    y = dfwork[ylbl].values
    r, _ = stats.spearmanr(x, y)
    res = dfwork.groupby('bin')[[xlbl, ylbl]].mean()
    
    ax.scatter(x,y,s=5, edgecolor='none') #, cmap='viridis')
    ax.plot(res[xlbl], res[ylbl], color='k')
    ax.set_title(f'r={r:.2f}')
    ax.set_xlabel(xlbl)
    ax.set_ylabel(ylbl)
    ax.set_xscale('log')
    ax.set_yscale('log')
fig.tight_layout()
plt.show()


dfwork = adata.obs.copy() 
xlbl = 'volume'
ylbl = 'jnorm_transcript_count'

x = dfwork[xlbl].values 
y = dfwork[ylbl].values
r, _ = stats.spearmanr(x, y)
bins = np.logspace(2,3.2,50)
dfwork['bin'] = pd.cut(x, bins)
res = dfwork.groupby('bin')[[xlbl, ylbl]].mean()

plt.scatter(x,y,s=5, edgecolor='none') #, cmap='viridis')
plt.plot(res[xlbl], res[ylbl], color='k')
plt.title(f'r={r:.2f}')
plt.xlabel(xlbl)
plt.ylabel(ylbl)
plt.xscale('log')
plt.yscale('log')

fig, axs = plt.subplots(1,2,figsize=(2*5,1*4), sharex=True, sharey=True)
ax = axs[0]
sns.boxplot(data=adata.obs, y='transcript_count', x='sample', ax=ax) #].values
ax.set_xticklabels(ax.get_xticklabels(), rotation=90)
ax.set_yscale('log')

ax = axs[1]
sns.boxplot(data=adata.obs, y='jnorm_transcript_count', x='sample', ax=ax) #].values
ax.set_xticklabels(ax.get_xticklabels(), rotation=90)
ax.set_yscale('log')

fig.tight_layout()
plt.show()


adata

# focus on Glu neurons only

r = 0.3
adata.obs['annot'] = clst_annots[adata.obs[f'leiden_r{r}'].astype(int)]
print(adata.obs['annot'].value_counts())

adata = adata[adata.obs['annot'].str.contains('Glu')]
print(adata)
# # adata = adata[adata.obs['gncov']>100]
# print(adata)

width_min = adata.obs.groupby('sample')['width'].min().reindex(names)
width_max = adata.obs.groupby('sample')['width'].max().reindex(names)
width_rng = width_max - width_min 
width_cum = pd.Series(np.cumsum(np.hstack([0, width_rng[:-1]+100])), index=names)

adata.obs['width_n0'] = adata.obs['width'] - width_min.reindex(adata.obs['sample']).values
adata.obs['width_show2'] =  adata.obs['width_n0'] + width_cum.reindex(adata.obs['sample']).values
adata.obs['depth_show2'] = -adata.obs['depth']

adata_nr = adata[adata.obs['sample'].str.contains('NR')].copy()
adata_dr = adata[adata.obs['sample'].str.contains('DR')].copy()

mat    = stats.zscore(adata[:,abcgenes].layers['ljnorm'], axis=0)
mat_nr = stats.zscore(adata_nr[:,abcgenes].layers['ljnorm'], axis=0)

pca = PCA(n_components=5)
adata.obsm['pcs_typegenes']    = pca.fit_transform(mat)
adata.obsm['pcs_typegenes_nr'] = pca.fit(mat_nr).transform(mat)
adata_nr.obsm['pcs_typegenes'] = pca.fit_transform(mat_nr)

xi, yi = 0, 1
xsign, ysign = 1, 1

metrics = ['gncov', 'gnnum', 'depth', 'width_show']

fig, axs = plt.subplots(1,4,figsize=(4*5,1*4))
for metric, ax in zip(metrics, axs):
    # g = np.log10(1+adata.obs[metric])
    g = adata.obs[metric]
    x = xsign*adata.obsm['pcs_typegenes'][:,xi]
    y = ysign*adata.obsm['pcs_typegenes'][:,yi]
    utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=3, )
    ax.set_title(metric)

xi, yi = 1, 2
xsign, ysign = -1,-1

%%time

from py_pcha import PCHA

np.random.seed(0)

X = np.vstack([
    xsign*adata.obsm['pcs_typegenes'][:,xi], 
    ysign*adata.obsm['pcs_typegenes'][:,yi], 
])

XC, S, C, SSE, varexpl = PCHA(X, noc=3, delta=0)
XC = np.array(XC)
XC = XC[:,np.argsort(XC[0])].copy() # order this
print(XC.shape, S.shape, C.shape, SSE.shape, varexpl.shape, SSE, varexpl)




def add_triangle(XC, ax, zorder=0, vertices=False, linecolor='gray', linewidth=1, **kwargs):
    # add the triangle
    ax.plot(XC[0].tolist()+[XC[0,0]], XC[1].tolist()+[XC[1,0]], '--', color=linecolor, label='', zorder=zorder, linewidth=linewidth, markersize=3)
    
    # add vertices
    if vertices:
        ax.scatter(XC[0,0], XC[1,0], color='C0', zorder=zorder, **kwargs)
        ax.scatter(XC[0,1], XC[1,1], color='C1', zorder=zorder, **kwargs)
        ax.scatter(XC[0,2], XC[1,2], color='C2', zorder=zorder, **kwargs)

gns = [agenes, bgenes, cgenes, iegs]
titles = ['A genes', 'B genes', 'C genes', 'IEGs']
adatas = [
    adata,
    adata[adata.obs['sample'].str.contains('NR')],
    adata[adata.obs['sample'].str.contains('DR')],
]
conditions = ['combined', 'NR', 'DR']

fig, axss = plt.subplots(3,4,figsize=(4*5,3*3), sharex=True, sharey=True)
for i, (axs, adatasub, condition) in enumerate(zip(axss, adatas, conditions)):
    condition = conditions[i]
    for j, (ax, gn, title,) in enumerate(zip(axs, gns, titles)):
        g = adatasub[:,gn].layers['ljnorm'].mean(axis=1)
        x = xsign*adatasub.obsm['pcs_typegenes'][:,xi]
        y = ysign*adatasub.obsm['pcs_typegenes'][:,yi]
        
        p = utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=5, cmap='coolwarm', vmin_p=5, vmax_p=95)
        colorbar = plt.colorbar(p, aspect=5, shrink=0.3)
        
        # Show ticks but no grid
        ax.set_aspect('equal')
        ax.axis('on')
        ax.grid(False)  # Turn off grid lines
        sns.despine(ax=ax)
        ax.tick_params(axis='both', which='both', bottom=True, left=True)
        
        if i == 0:
            ax.set_title(title)
        if j == 0:
            ax.set_ylabel(condition, rotation=0, loc='top')
            
        # add the triangle
        add_triangle(XC, ax)
plt.show()

gns = [agenes, bgenes, cgenes, iegs]
titles = ['A genes', 'B genes', 'C genes', 'IEGs']
adatas = [
    adata,
    adata[adata.obs['sample'].str.contains('NR')],
    adata[adata.obs['sample'].str.contains('DR')],
]
conditions = ['combined', 'NR', 'DR']

fig, axss = plt.subplots(3,4,figsize=(4*5,3*3), sharex=True, sharey=True)
for i, (axs, adatasub, condition) in enumerate(zip(axss, adatas, conditions)):
    condition = conditions[i]
    for j, (ax, gn, title,) in enumerate(zip(axs, gns, titles)):
        g = adatasub[:,gn].layers['ljnorm'].mean(axis=1)
        x = xsign*adatasub.obsm['pcs_typegenes'][:,xi]
        y = ysign*adatasub.obsm['pcs_typegenes'][:,yi]
        
        # consistent over
        g0 = adata[:,gn].layers['ljnorm'].mean(axis=1)
        vmin = np.percentile(g0,  5)
        vmax = np.percentile(g0, 95)
            
        p = utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=5, cmap='coolwarm', vmin=vmin, vmax=vmax)
        colorbar = plt.colorbar(p, aspect=5, shrink=0.3)
        
        # Show ticks but no grid
        ax.set_aspect('equal')
        ax.axis('on')
        ax.grid(False)  # Turn off grid lines
        sns.despine(ax=ax)
        ax.tick_params(axis='both', which='both', bottom=True, left=True)
        
        if i == 0:
            ax.set_title(title)
        if j == 0:
            ax.set_ylabel(condition, rotation=0, loc='top')
            
        # add the triangle
        add_triangle(XC, ax)
plt.show()

gns = ['Cdh13', 'Sorcs3', 'Chrm2', 'Fos']
titles = gns
adatas = [
    adata,
    adata[adata.obs['sample'].str.contains('NR')],
    adata[adata.obs['sample'].str.contains('DR')],
]
conditions = ['combined', 'NR', 'DR']

fig, axss = plt.subplots(3,4,figsize=(4*5,3*3), sharex=True, sharey=True)
for i, (axs, adatasub, condition) in enumerate(zip(axss, adatas, conditions)):
    condition = conditions[i]
    for j, (ax, gn, title,) in enumerate(zip(axs, gns, titles)):
        g = adatasub[:,gn].layers['ljnorm'].reshape(-1,)
        x = xsign*adatasub.obsm['pcs_typegenes'][:,xi]
        y = ysign*adatasub.obsm['pcs_typegenes'][:,yi]
        
        # consistent over
        g0 = adata[:,gn].layers['ljnorm'].mean(axis=1)
        vmin = np.percentile(g0,  5)
        vmax = np.percentile(g0, 99)
            
        p = utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=5, vmin=vmin, vmax=vmax)
        colorbar = plt.colorbar(p, aspect=5, shrink=0.3)
        
        # Show ticks but no grid
        ax.set_aspect('equal')
        ax.axis('on')
        ax.grid(False)  # Turn off grid lines
        sns.despine(ax=ax)
        ax.tick_params(axis='both', which='both', bottom=True, left=True)
        
        if i == 0:
            ax.set_title(title)
        if j == 0:
            ax.set_ylabel(condition, rotation=0, loc='top')
            
        add_triangle(XC, ax)
plt.show()

gns = ['Syt17', 'Astn2', 'Etl4', 'Slc24a3'] 
titles = gns
adatas = [
    adata,
    adata[adata.obs['sample'].str.contains('NR')],
    adata[adata.obs['sample'].str.contains('DR')],
]
conditions = ['combined', 'NR', 'DR']

fig, axss = plt.subplots(3,4,figsize=(4*5,3*3), sharex=True, sharey=True)
for i, (axs, adatasub, condition) in enumerate(zip(axss, adatas, conditions)):
    condition = conditions[i]
    for j, (ax, gn, title,) in enumerate(zip(axs, gns, titles)):
        g = adatasub[:,gn].layers['ljnorm'].reshape(-1,)
        x = xsign*adatasub.obsm['pcs_typegenes'][:,xi]
        y = ysign*adatasub.obsm['pcs_typegenes'][:,yi]
        
        # consistent over
        g0 = adata[:,gn].layers['ljnorm'].mean(axis=1)
        vmin = np.percentile(g0,  5)
        vmax = np.percentile(g0, 99)
            
        p = utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=5, vmin=vmin, vmax=vmax)
        colorbar = plt.colorbar(p, aspect=5, shrink=0.3)
        
        # Show ticks but no grid
        ax.set_aspect('equal')
        ax.axis('on')
        ax.grid(False)  # Turn off grid lines
        sns.despine(ax=ax)
        ax.tick_params(axis='both', which='both', bottom=True, left=True)
        
        if i == 0:
            ax.set_title(title)
        if j == 0:
            ax.set_ylabel(condition, rotation=0, loc='top')
            
        add_triangle(XC, ax)
plt.show()

# heatmap type-identity genes vs cells
- gene order from snRNA-seq
- cell order from pseudotime

ordered_typegenes = typegenes_df.sort_values('gene_ptime')['gene_name_vizgen'].values
print(ordered_typegenes.shape)
ordered_typegenes

def gen_ptime(adata, use_rep='X_pca', sign=+1):
    """
    """
    pcs = adata.obsm[use_rep]
    
    sc.pp.neighbors(adata, n_neighbors=50, use_rep=use_rep)
    sc.tl.diffmap(adata)
    adata.uns['iroot'] = np.argmin(sign*pcs[:,0])
    sc.tl.dpt(adata)
    cell_ptime = adata.obs['dpt_pseudotime'].values
    return cell_ptime

pc0 = adata_nr.obsm['pcs_typegenes'][:,0]
depth = adata_nr.obs['depth'] #   = gen_ptime(adata_nr, use_rep='ptime_rep', sign=-1)

cond0 = depth > 100
cond1 = pc0 > np.percentile(pc0, 5)
# cond2 = depth < 350
adata_nr2 = adata_nr[cond0 & cond1]
# adata_nr2 = adata_nr[cond0 & cond1 & cond2]
print(adata_nr2)

plt.scatter(depth, pc0, s=2)

adata_nr2.obsm['ptime_rep'] = adata_nr2.obsm['pcs_typegenes'][:,[1,2]]
cell_ptime = gen_ptime(adata_nr2, use_rep='ptime_rep', sign=-1)
depth = adata_nr2.obs['depth'] #   = gen_ptime(adata_nr, use_rep='ptime_rep', sign=-1)
plt.scatter(depth, cell_ptime, s=2)

adata_nr_ordered = adata_nr2[np.argsort(cell_ptime), ordered_typegenes].copy()
adata_nr_ordered2 = adata_nr2[np.argsort(depth), ordered_typegenes].copy()
adata_nr_ordered

bigmat = stats.zscore(adata_nr_ordered.layers['ljnorm'],axis=0)
bigmat2 = stats.zscore(adata_nr_ordered2.layers['ljnorm'],axis=0)
bigdepth = adata_nr_ordered.obs['depth']
bigdepth2 = adata_nr_ordered2.obs['depth']

r, _ = stats.spearmanr(depth, cell_ptime)
plt.title(f'Spearman r = {r:.2f}')

bins = np.linspace(100,400,20)
tmp = pd.DataFrame()
tmp['bin'] = pd.cut(depth, bins)
tmp['ptime'] = cell_ptime
mean_ptime = tmp.groupby('bin').mean()['ptime']

plt.scatter(depth, cell_ptime, s=1)
plt.plot(bins[1:], mean_ptime, color='k')


bigdepth.max()


mosaic = "A\n"*20+"B"

fig, axs = plt.subplot_mosaic(mosaic, figsize=(12,6))
ax = axs['A']
sns.heatmap(bigmat.T,
            xticklabels=False,
            yticklabels=False,
            cbar_kws=dict(shrink=0.3, label='zscore log10CP10k', aspect=10),
            vmax=3,
            vmin=-3,
            cmap='coolwarm',
            rasterized=True,
            ax=ax,
           )

ax = axs['B']
sns.heatmap(bigdepth.values.reshape(-1,1).T, 
            xticklabels=False, 
            yticklabels=False, 
            cmap='gray_r', 
            vmin=100, vmax=400,
            cbar_kws=dict(shrink=2, aspect=5, ticks=[100,400]),
            ax=ax)
plt.show()

mosaic = "A\n"*20+"B"

fig, axs = plt.subplot_mosaic(mosaic, figsize=(12,6))
ax = axs['A']
sns.heatmap(bigmat2.T,
            xticklabels=False,
            yticklabels=False,
            cbar_kws=dict(shrink=0.3, label='zscore log10CP10k', aspect=10),
            vmax=3,
            vmin=-3,
            cmap='coolwarm',
            rasterized=True,
            ax=ax,
           )

ax = axs['B']
sns.heatmap(bigdepth2.values.reshape(-1,1).T, 
            xticklabels=False, 
            yticklabels=False, 
            cmap='gray_r', 
            vmin=100, vmax=400,
            cbar_kws=dict(shrink=2, aspect=5, ticks=[100,400]),
            ax=ax)
plt.show()


adatasub = adata_nr2

fig, ax = plt.subplots()

g = cell_ptime#adatasub[:,gn].layers['ljnorm'].reshape(-1,)
x = xsign*adatasub.obsm['pcs_typegenes'][:,xi]
y = -ysign*adatasub.obsm['pcs_typegenes'][:,yi]

# consistent over
p = utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=5, cmap='coolwarm') # vmin=vmin, vmax=vmax)
colorbar = plt.colorbar(p, aspect=5, shrink=0.3)

# Show ticks but no grid
ax.set_aspect('equal')
ax.axis('on')
ax.grid(False)  # Turn off grid lines
sns.despine(ax=ax)
ax.tick_params(axis='both', which='both', bottom=True, left=True)

if i == 0:
    ax.set_title(title)
if j == 0:
    ax.set_ylabel(condition, rotation=0, loc='top')

add_triangle(XC, ax)

# customized colormap 

from matplotlib.colors import LinearSegmentedColormap

colors_a = [(0.0, 'black'), (1.0, 'C0')]      
colors_b = [(0.0, 'black'), (1.0, 'C1')]      
colors_c = [(0.0, 'black'), (1.0, 'C2')]      
colors_nrdr = [(0.0, 'C1'), (0.5, 'white'), (1.0, 'black')]
colors_nr = [(0.0, 'white'), (1.0, 'C1'),]
colors_dr = [(0.0, 'white'), (1.0, 'black'),]

# Create a custom colormap using LinearSegmentedColormap
cmap_a = LinearSegmentedColormap.from_list('cmap_a', colors_a)
cmap_b = LinearSegmentedColormap.from_list('cmap_b', colors_b)
cmap_c = LinearSegmentedColormap.from_list('cmap_c', colors_c)
cmap_nrdr = LinearSegmentedColormap.from_list('cmap_nrdr', colors_nrdr)
cmap_nr = LinearSegmentedColormap.from_list('cmap_nr', colors_nr)
cmap_dr = LinearSegmentedColormap.from_list('cmap_dr', colors_dr)

xmin, xmax = -12, 12
ymin, ymax = -7, 7 

bins_x = np.linspace(xmin, xmax, 1*(xmax-xmin)+1)
bins_y = np.linspace(ymin, ymax, 1*(ymax-ymin)+1)
print(bins_x)
print(bins_y)

hists = []
fig, axs = plt.subplots(1,4,figsize=(4*5,1*4), sharex=True, sharey=True)
for ax, adatasub, cond, _cmap in zip(axs, [
    adata,
    adata[adata.obs['sample'].str.contains('NR')],
    adata[adata.obs['sample'].str.contains('DR')],
    ], 
    ['Combined', 'NR', 'DR'], 
    ['gray_r', cmap_nr, cmap_dr]):
    x =  xsign*adatasub.obsm['pcs_typegenes'][:,xi]
    y =  ysign*adatasub.obsm['pcs_typegenes'][:,yi]
    sns.histplot(x=x, y=y, ax=ax, bins=(bins_x, bins_y), 
                 cmap=_cmap, # 'gray_r', 
                 stat='percent', vmin=0, vmax=2, 
                 cbar=True, cbar_kws=dict(shrink=0.4, ticks=[0,2]))
    # sns.kdeplot(x=x, y=y, ax=ax, bins=(bins_x, bins_y))
    
    hist, _, _= np.histogram2d(x, y, bins=(bins_x, bins_y))
    hist = hist/len(x)*100
    hists.append(hist)
    print(hist.shape)
    ax.set_title(cond)
    ax.set_aspect('equal')
    sns.despine(ax=ax)
    ax.grid(False)
    
    # g = ax.imshow(pd.DataFrame(np.log2(1e-3+hist), 
    #                          index=bins_x[1:]-0.5, 
    #                          columns=bins_y[1:]-0.5).T, 
    #             origin='lower',
    #             extent=(xmin, xmax, ymin, ymax),
    #             cmap='gray_r') # , vmax=1, vmin=-1)
    
    # add the triangle
    add_triangle(XC, ax, zorder=2)
    
ax = axs[3] 
ax.set_title('DR-NR')
g = ax.imshow(
    # pd.DataFrame(np.log2(1e-3+hists[2])-np.log2(1e-3+hists[1]), #-hists[1], 
    pd.DataFrame(hists[2]-hists[1], #-hists[1], 
                         index=bins_x[1:]-0.5, 
                         columns=bins_y[1:]-0.5).T, 
            origin='lower',
            extent=(xmin, xmax, ymin, ymax),
            # cmap='coolwarm', 
            cmap=cmap_nrdr, 
            vmax=2, vmin=-2)
# ax.invert_yaxis()
ax.set_aspect('equal')
ax.grid(False)
fig.colorbar(g, shrink=0.4, ticks=[-2,0,2])
sns.despine(ax=ax)

# add the triangle
add_triangle(XC, ax, zorder=2)

plt.show()

dfshow = adata.obs.copy()
dfshow['nrdr'] = dfshow['sample'].str.contains('DR').astype(int)
dfshow['dim1'] = xsign*adata.obsm['pcs_typegenes'][:,xi]
dfshow['dim2'] = ysign*adata.obsm['pcs_typegenes'][:,yi]
palette = {0: 'C1', 1: 'black'}


fig, axs = plt.subplots(1,2,figsize=(2*5,1*4), sharex=True, sharey=True) 
ax = axs[0]
add_triangle(XC, ax)
sns.scatterplot(data=dfshow.sample(frac=1), x='dim1', y='dim2', hue='nrdr', s=3, edgecolor='none', palette=palette, ax=ax)
ax.set_aspect('equal')
sns.despine(ax=ax)
ax.grid(False)
ax.legend(bbox_to_anchor=(1,1))

ax = axs[1]
add_triangle(XC, ax)
sns.kdeplot(data=dfshow, x='dim1', y='dim2', hue='nrdr', palette=palette, legend=False, ax=ax,)
ax.set_aspect('equal')
sns.despine(ax=ax)
ax.grid(False)


fig.tight_layout()
plt.show()

fig, axs = plt.subplots(2,4,figsize=(4*4,2*3), sharex=True, sharey=True) 
for sample, ax in zip(names, axs.flat):
    add_triangle(XC, ax)
    sns.scatterplot(data=dfshow[dfshow['sample']==sample].sample(frac=1), 
                    x='dim1', y='dim2', hue='nrdr', s=5, edgecolor='none', palette=palette, ax=ax, legend=False)
    sns.kdeplot(data=dfshow[dfshow['sample']==sample],
                x='dim1', y='dim2', hue='nrdr', palette=palette, legend=False, ax=ax,)
    ax.set_aspect('equal')
    sns.despine(ax=ax)
    ax.set_title(sample)
    ax.grid(False)
    # ax.legend(bbox_to_anchor=(1,1))

fig.tight_layout()
plt.show()

# customized colormap 

from matplotlib.colors import LinearSegmentedColormap

colors_a = [(0.0, 'black'), (1.0, 'C0')]      
colors_b = [(0.0, 'black'), (1.0, 'C1')]      
colors_c = [(0.0, 'black'), (1.0, 'C2')]      

# Create a custom colormap using LinearSegmentedColormap
cmap_a = LinearSegmentedColormap.from_list('cmap_a', colors_a)
cmap_b = LinearSegmentedColormap.from_list('cmap_b', colors_b)
cmap_c = LinearSegmentedColormap.from_list('cmap_c', colors_c)


def assign_types_and_colors(adata):
    """
    """
    # get ABC scores
    g0_a = adata[:,agenes].layers['ljnorm'].mean(axis=1)
    g0_b = adata[:,bgenes].layers['ljnorm'].mean(axis=1)
    g0_c = adata[:,cgenes].layers['ljnorm'].mean(axis=1)

    # make ABC scores comparable and norm to [0,1] [50% to 95%]
    vmin_p, vmax_p = 30, 95
    vmin_a = np.percentile(g0_a, vmin_p)
    vmax_a = np.percentile(g0_a, vmax_p)

    vmin_b = np.percentile(g0_b, vmin_p)
    vmax_b = np.percentile(g0_b, vmax_p)

    vmin_c = np.percentile(g0_c, vmin_p)
    vmax_c = np.percentile(g0_c, vmax_p)

    g0_a = np.clip((g0_a-vmin_a)/(vmax_a-vmin_a), 0, 1)
    g0_b = np.clip((g0_b-vmin_b)/(vmax_b-vmin_b), 0, 1)
    g0_c = np.clip((g0_c-vmin_c)/(vmax_c-vmin_c), 0, 1)

    # separate them into scale and frequency
    g0_sum  = (g0_a+g0_b+g0_c)
    g0_size = np.clip(g0_sum/np.percentile(g0_sum, 50), 0, 1) # [0 - > 50%] - [0, 1]
    freq0_a = g0_a/(g0_sum+1e-5)
    freq0_b = g0_b/(g0_sum+1e-5)
    freq0_c = g0_c/(g0_sum+1e-5)

    adata.obs['raw_score_a'] = adata[:,agenes].layers['ljnorm'].mean(axis=1)
    adata.obs['raw_score_b'] = adata[:,bgenes].layers['ljnorm'].mean(axis=1)
    adata.obs['raw_score_c'] = adata[:,cgenes].layers['ljnorm'].mean(axis=1)
    adata.obsm['size_freq_abc'] = np.vstack([g0_size, freq0_a, freq0_b, freq0_c, g0_sum]).T
    adata.obsm['additive_rgb'] = (cmap_a(freq0_a)+cmap_b(freq0_b)+cmap_c(freq0_c))[:,:3]
    
    return

assign_types_and_colors(adata)
assign_types_and_colors(adata_nr)
assign_types_and_colors(adata_dr)

# ABC scores - expression level distributions

ascores_nr = -adata[adata.obs['sample'].str.contains(f'NR')][:,agenes].layers['ljnorm'].mean(axis=1)
ascores_dr = -adata[adata.obs['sample'].str.contains(f'DR')][:,agenes].layers['ljnorm'].mean(axis=1)

bscores_nr = -adata[adata.obs['sample'].str.contains(f'NR')][:,bgenes].layers['ljnorm'].mean(axis=1)
bscores_dr = -adata[adata.obs['sample'].str.contains(f'DR')][:,bgenes].layers['ljnorm'].mean(axis=1)

cscores_nr = -adata[adata.obs['sample'].str.contains(f'NR')][:,cgenes].layers['ljnorm'].mean(axis=1)
cscores_dr = -adata[adata.obs['sample'].str.contains(f'DR')][:,cgenes].layers['ljnorm'].mean(axis=1)

fig, axs = plt.subplots(1,3, figsize=(3*4,4), sharex=False, sharey=False)

ax = axs[0]
sns.histplot(ascores_nr, ax=ax, stat='percent', cumulative=True, element='step', fill=False, color='C1')
sns.histplot(ascores_dr, ax=ax, stat='percent', cumulative=True, element='step', fill=False, color='black')
ax.set_title('A score')
sns.despine(ax=ax)
ax.grid(False)

ax = axs[1]
sns.histplot(bscores_nr, ax=ax, stat='percent', cumulative=True, element='step', fill=False, color='C1')
sns.histplot(bscores_dr, ax=ax, stat='percent', cumulative=True, element='step', fill=False, color='black')
ax.set_title('B score')
sns.despine(ax=ax)
ax.grid(False)

ax = axs[2]
sns.histplot(cscores_nr, ax=ax, stat='percent', cumulative=True, element='step', fill=False, color='C1')
sns.histplot(cscores_dr, ax=ax, stat='percent', cumulative=True, element='step', fill=False, color='black')
ax.set_title('C score')
sns.despine(ax=ax)
ax.grid(False)

fig.tight_layout()
plt.show()

fig, axs = plt.subplots(1,4, figsize=(4*4,4), sharex=False, sharey=True)
for ax, genegroup, title in zip(axs, 
                                [agenes, bgenes, cgenes, iegs], 
                                ['A genes', 'B genes', 'C genes', 'IEGs'],
                               ):
    for i, sample in enumerate(names):
        scores_ = adata[adata.obs['sample']==sample][:,genegroup].layers['ljnorm'].mean(axis=1)
        if 'NR' in sample:
            color = 'C1'
        elif 'DR' in sample:
            color = 'black'

        sns.ecdfplot(scores_, ax=ax, color=color)#, complementary=True) # , linewidth=2)
        
        # _x = np.percentile(scores_, 50)
        # _y = 0.5
        # ax.text(_x, _y, i, fontsize=10, color='red')
    ax.set_ylabel('Cumulative proportion\nof cells')
    ax.set_xlabel('Mean log norm expr.')
    ax.set_title(title)
    sns.despine(ax=ax)
    ax.grid(False)
    

fig.tight_layout()
plt.show()

xi, yi = 1,2
xsign, ysign = -1,-1

adatas = [
    adata, 
    adata[adata_nr.obs.index], 
    adata[adata_dr.obs.index], 
]
conditions = ['combined', 'NR', 'DR']

fig, axs = plt.subplots(1,3,figsize=(3*5,1*3), sharex=True, sharey=True)
for i, (ax, adatasub, condition) in enumerate(zip(axs, adatas, conditions)):
    condition = conditions[i]
    x = xsign*adatasub.obsm['pcs_typegenes'][:,xi]
    y = ysign*adatasub.obsm['pcs_typegenes'][:,yi]
    
    g_size = adatasub.obsm['size_freq_abc'][:,0]
    freq_a = adatasub.obsm['size_freq_abc'][:,1]
    freq_b = adatasub.obsm['size_freq_abc'][:,2]
    freq_c = adatasub.obsm['size_freq_abc'][:,3]
    rgbs   = adatasub.obsm['additive_rgb']
    
    # visualize ABC scores using additive blending
    p = ax.scatter(x, y, c=rgbs, s=5, edgecolor='none') #, alpha=g_size)
        
    # Show ticks but no grid
    ax.set_aspect('equal')
    ax.axis('on')
    ax.grid(False)  # Turn off grid lines
    sns.despine(ax=ax)
    ax.tick_params(axis='both', which='both', bottom=True, left=True)
    ax.set_title(condition)

    # add the triangle
    add_triangle(XC, ax, vertices=True, edgecolors='k', linewidths=1, marker='o')
    # break
# plt.show()

res = []
for sample in names:
    if 'NR' in sample:
        cond = 'NR'
    elif 'DR' in sample:
        cond = 'DR'
        
    adatasub = adata[adata.obs['sample']==sample]
    g_size = adatasub.obsm['size_freq_abc'][:,0]
    g_sum  = adatasub.obsm['size_freq_abc'][:,4]
    freq_a = adatasub.obsm['size_freq_abc'][:,1]
    freq_b = adatasub.obsm['size_freq_abc'][:,2]
    freq_c = adatasub.obsm['size_freq_abc'][:,3]

    # ta = np.sum(g_size*(freq_a))
    # tb = np.sum(g_size*(freq_b))
    # tc = np.sum(g_size*(freq_c))
    ta = np.sum(g_sum*(freq_a))
    tb = np.sum(g_sum*(freq_b))
    tc = np.sum(g_sum*(freq_c))
    tt = (ta+tb+tc)
    # print(sample, f'{ta/tt: .2f}', f'{tb/tt: .2f}', f'{tc/tt: .2f}')
    res.append([sample, cond, ta/tt*100, tb/tt*100, tc/tt*100, ta, tb, tc, tt])
    
res = pd.DataFrame(res, columns=['sample', 'cond', 'A', 'B', 'C', 'tA', 'tB', 'tC', 'tt']).set_index('sample')
res

for i in range(3):
    print(stats.ttest_ind(res.iloc[:4,1+i], res.iloc[4:,1+i]))

fig, axs = plt.subplots(1, 2, figsize=(4,3), sharey=True)
ax = axs[0]
res.loc[names[:4]].plot.bar(ax=ax, stacked=True, width=0.8, edgecolor='none', legend=False)
ax.grid(False)
sns.despine(ax=ax)
ax.set_title('NR')
ax.set_ylabel('L2/3 cells (%)')

ax = axs[1]
res.loc[names[4:]].plot.bar(ax=ax, stacked=True, width=0.8, edgecolor='none')
ax.legend(bbox_to_anchor=(1,1))
ax.grid(False)
sns.despine(ax=ax)
ax.set_title('DR')

plt.show()

for i, lbl in enumerate(['A', 'B', 'C']):
    fig, axs = plt.subplots(1, 2, figsize=(4,3), sharey=True)
    ax = axs[0]
    res.loc[names[:4]][lbl].plot.bar(ax=ax, stacked=False, width=0.8, edgecolor='none', color=f'C{i}', legend=False)
    ax.grid(False)
    sns.despine(ax=ax)
    ax.set_title('NR')
    ax.set_ylabel('L2/3 cells (%)')

    ax = axs[1]
    res.loc[names[4:]][lbl].plot.bar(ax=ax, stacked=False, width=0.8, edgecolor='none', color=f'C{i}')
    ax.grid(False)
    sns.despine(ax=ax, left=True)
    ax.set_title('DR')


    plt.show()

from statsmodels.stats.proportion import test_proportions_2indep

fig, axs = plt.subplots(1, 3, figsize=(2*3,4))
for ax, col, color in zip(axs, ['A', 'B', 'C'], ['C0', 'C1', 'C2']):
    sns.barplot(data=res, x='cond', y=col, ax=ax, color=color, capsize=0.3, errwidth=1)
    sns.swarmplot(data=res, x='cond', y=col, color='k', ax=ax, )
    ax.set_title(f'type {col}', y=1.1)
    sns.despine(ax=ax)
    ax.grid(False)
    ax.set_ylabel('')
    
    a = res[res['cond']=='NR'][col]/100
    b = res[res['cond']=='DR'][col]/100
    # t, p = stats.ttest_ind(a, b)
    s, p = stats.mannwhitneyu(a, b)
    # testres = test_proportions_2indep(np.mean(a),4,np.mean(b),4)
    
    if p > 0.05:
        mark = 'ns'
    elif p < 0.05 and p > 0.001:
        mark = '*'
    elif p < 0.001:
        mark = '***'
    
    # statistical annotation
    x1, x2 = 0, 1   
    y, h = res[col].max() + 2, 2
    ax.plot([x1, x1, x2, x2], [y, y+h, y+h, y], lw=1.5, c='k')
    ax.text((x1+x2)*.5, y+h, mark, ha='center', va='bottom', color='k')
    
axs[0].set_ylabel('L2/3 cells (%)')
fig.tight_layout()
plt.show()

fig, axs = plt.subplots(1, 3, figsize=(2*3,4))
for ax, col, color in zip(axs, ['A', 'B', 'C'], ['C0', 'C1', 'C2']):
    sns.boxplot(data=res, x='cond', y=col, color='white', ax=ax)
    sns.swarmplot(data=res, x='cond', y=col, color=color, ax=ax)
    ax.set_title(f'type {col}')
    sns.despine(ax=ax)
    ax.grid(False)
    ax.set_ylabel('')
    
axs[0].set_ylabel('L2/3 cells (%)')
fig.tight_layout()
plt.show()

res = []
for sample in names:
    if 'NR' in sample:
        cond = 'NR'
    elif 'DR' in sample:
        cond = 'DR'
        
    adatasub = adata[adata.obs['sample']==sample]
    depths = adatasub.obs['depth'].values
    g_size = adatasub.obsm['size_freq_abc'][:,0]
    g_sum  = adatasub.obsm['size_freq_abc'][:,4]
    freq_a = adatasub.obsm['size_freq_abc'][:,1]
    freq_b = adatasub.obsm['size_freq_abc'][:,2]
    freq_c = adatasub.obsm['size_freq_abc'][:,3]

    # ta = np.sum(g_size*(freq_a))
    # tb = np.sum(g_size*(freq_b))
    # tc = np.sum(g_size*(freq_c))
    ta = np.sum(g_sum*(freq_a))
    tb = np.sum(g_sum*(freq_b))
    tc = np.sum(g_sum*(freq_c))
    tt = (ta+tb+tc)
    print(sample, f'{ta/tt: .2f}', f'{tb/tt: .2f}', f'{tc/tt: .2f}')
    res.append([sample, cond, ta/tt*100, tb/tt*100, tc/tt*100])
    
res = pd.DataFrame(res, columns=['sample', 'cond', 'A', 'B', 'C']).set_index('sample')

for sample in names:
    adatasub = adata[adata.obs['sample']==sample]
    g_size = adatasub.obsm['size_freq_abc'][:,0]
    g_sum  = adatasub.obsm['size_freq_abc'][:,4]
    freq_a = adatasub.obsm['size_freq_abc'][:,1]
    freq_b = adatasub.obsm['size_freq_abc'][:,2]
    freq_c = adatasub.obsm['size_freq_abc'][:,3]

    # ta = np.sum(g_size*(freq_a))
    # tb = np.sum(g_size*(freq_b))
    # tc = np.sum(g_size*(freq_c))
    ta = np.sum(g_size*(freq_a))
    tb = np.sum(g_size*(freq_b))
    tc = np.sum(g_size*(freq_c))
    tt = (ta+tb+tc)
    print(sample, f'{ta/tt: .2f}', f'{tb/tt: .2f}', f'{tc/tt: .2f}')

for sample in names:
    adatasub = adata[adata.obs['sample']==sample]
    g_size = adatasub.obsm['size_freq_abc'][:,0]
    g_sum  = adatasub.obsm['size_freq_abc'][:,4]
    freq_a = adatasub.obsm['size_freq_abc'][:,1]
    freq_b = adatasub.obsm['size_freq_abc'][:,2]
    freq_c = adatasub.obsm['size_freq_abc'][:,3]

    rank = np.argsort(np.vstack([freq_a,freq_b,freq_c]).T, axis=1)[:,-1]
    ta = np.sum(rank==0)
    tb = np.sum(rank==1)
    tc = np.sum(rank==2)
    tt = len(g_size)
    print(sample, f'{ta/tt: .2f}', f'{tb/tt: .2f}', f'{tc/tt: .2f}')

x =  adata.obs['width_show2']
y =  adata.obs['depth_show2']

g_size = adata.obsm['size_freq_abc'][:,0]
freq_a = adata.obsm['size_freq_abc'][:,1]
freq_b = adata.obsm['size_freq_abc'][:,2]
freq_c = adata.obsm['size_freq_abc'][:,3]

# visualize ABC scores using additive blending
additive = (cmap_a(freq_a)+cmap_b(freq_b)+cmap_c(freq_c))[:,:3]

fig, ax = plt.subplots(1,1,figsize=(1*25,1))
for lbl, coord in width_cum.items():
    ax.text(coord, 0, lbl, fontsize=12)
    
sorting = np.argsort(np.max(additive, axis=1))# [::-1]
p = ax.scatter(x[sorting], y[sorting], c=additive[sorting], s=5, edgecolor='none', alpha=g_size[sorting])
# cond = np.max(additive, axis=1) > 1e-5
# p = ax.scatter(x[cond], y[cond], c=normed[cond], s=1, edgecolor='none')
ax.set_aspect('equal')
ax.axis('off')

plt.show()

# visualize FISH

gns = ['Cdh13', 'Sorcs3', 'Chrm2', 'Fos'] 
x =  adata.obs['width_show2']
y =  adata.obs['depth_show2']
n = len(gns)

fig, axs = plt.subplots(n,1,figsize=(1*25,n*1))
for i, (ax, gn) in enumerate(zip(axs, gns)):
    if i == 0:
        for lbl, coord in width_cum.items():
            ax.text(coord, 0, lbl, fontsize=12)
    
    g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))
    vmax = np.percentile(g, 99)
    vmin = np.percentile(g,  5)
    sorting = np.argsort(g)
    
    p = utils_merfish.st_scatter_ax(fig, ax,  x[sorting],  y[sorting],  gexp=g[sorting], s=5, title='', vmin=vmin, vmax=vmax, cmap='rocket_r')
    ax.set_title(gn, loc='left', va='center', ha='right', y=0.5, pad=None)
    fig.colorbar(p, pad=0, shrink=0.5, aspect=5, ticks=[np.round(vmin, decimals=1), np.round(vmax-0.1, decimals=1)])
    
plt.show()
    

gns = ['Syt17', 'Astn2', 'Etl4', 'Slc24a3'] 
x =  adata.obs['width_show2']
y =  adata.obs['depth_show2']
n = len(gns)

fig, axs = plt.subplots(n,1,figsize=(1*25,n*1))
for i, (ax, gn) in enumerate(zip(axs, gns)):
    if i == 0:
        for lbl, coord in width_cum.items():
            ax.text(coord, 0, lbl, fontsize=12)
    
    g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))
    vmax = np.percentile(g, 99)
    vmin = np.percentile(g,  5)
    sorting = np.argsort(g)
    
    p = utils_merfish.st_scatter_ax(fig, ax,  x[sorting],  y[sorting],  gexp=g[sorting], s=5, title='', vmin=vmin, vmax=vmax, cmap='rocket_r')
    ax.set_title(gn, loc='left', va='center', ha='right', y=0.5, pad=None)
    fig.colorbar(p, pad=0, shrink=0.5, aspect=5, ticks=[np.round(vmin, decimals=1), np.round(vmax-0.1, decimals=1)])
    
plt.show()
    

x =  adata.obs['width_show2']
y =  adata.obs['depth_show2']
gns = [agenes, bgenes, cgenes, iegs] 
titles = ['A genes', 'B genes', 'C genes', 'IEGs']
n = len(gns)

fig, axs = plt.subplots(n,1,figsize=(1*25,n*1))
for i, (ax, gn, title) in enumerate(zip(axs, gns, titles)):
    if i == 0:
        for lbl, coord in width_cum.items():
            ax.text(coord, 0, lbl, fontsize=12)
    
    g = adata[:,gn].layers['ljnorm'].mean(axis=1)
    sorting = np.argsort(g)
    
    vmin = np.percentile(g,  5)
    vmax = np.percentile(g, 95)
    p = utils_merfish.st_scatter_ax(fig, ax, x[sorting], y[sorting], gexp=g[sorting], 
                                    s=5, title='', vmin=vmin, vmax=vmax, cmap='coolwarm') #, axis_off=False)
    ax.set_title(title, loc='left', va='center', ha='right', y=0.5, pad=None)
    fig.colorbar(p, pad=0, shrink=0.5, aspect=5, ticks=[np.round(vmin, decimals=1), np.round(vmax-0.1, decimals=1)])
    
plt.show()
    

# stats

stats = {}
bins = np.linspace(0, 400, 12)

for name in names:
    adatasub = adata[adata.obs['sample']==name]# v1l23_data[name]
    lnorm_mean, lnorm_sem, lnorm_std, n, d, db = binning_pipe2(adatasub, 'depth', 'ljnorm', bins=bins)
    stats[name] = (lnorm_mean, lnorm_sem, lnorm_std, n, d, db)
d.value_counts()

# mean expression level across V1 L2/3 in NR
base_a0 = []
base_b0 = []
base_c0 = []
base_i0 = []
for name in [
    'P28NR_ant', 'P28NR_pos', 'P28NRb_ant', 'P28NRb_pos',]:
    (lnorm_mean, lnorm_sem, lnorm_std, n, d, db) = stats[name]
    base_a = np.mean(lnorm_mean.iloc[:,agenes_idx], axis=0) # across depth bins for each gene
    base_b = np.mean(lnorm_mean.iloc[:,bgenes_idx], axis=0) # across depth bins for each gene
    base_c = np.mean(lnorm_mean.iloc[:,cgenes_idx], axis=0) # across depth bins for each gene
    base_i = np.mean(lnorm_mean.iloc[:,igenes_idx], axis=0) # across depth bins for each gene
    
    base_a0.append(base_a)
    base_b0.append(base_b)
    base_c0.append(base_c)
    base_i0.append(base_i)
    
base_a0 = np.mean(base_a0, axis=0)
base_b0 = np.mean(base_b0, axis=0)
base_c0 = np.mean(base_c0, axis=0)
base_i0 = np.mean(base_i0, axis=0)

base_a0.shape, base_b0.shape, base_c0.shape, base_i0.shape

means = {}
sems = {}
for name in names:
    (lnorm_mean, lnorm_sem, lnorm_std, n, d, db) = stats[name]
    
    amean = np.mean(lnorm_mean.iloc[:,agenes_idx]-base_a0, axis=1) # a bin vector
    bmean = np.mean(lnorm_mean.iloc[:,bgenes_idx]-base_b0, axis=1) # a bin vector
    cmean = np.mean(lnorm_mean.iloc[:,cgenes_idx]-base_c0, axis=1) # a bin vector
    imean = np.mean(lnorm_mean.iloc[:,igenes_idx]-base_i0, axis=1) # a bin vector
    
    asem = np.mean(lnorm_sem.iloc[:,agenes_idx], axis=1)
    bsem = np.mean(lnorm_sem.iloc[:,bgenes_idx], axis=1)
    csem = np.mean(lnorm_sem.iloc[:,cgenes_idx], axis=1)
    isem = np.mean(lnorm_sem.iloc[:,igenes_idx], axis=1)
    
    means[name] = [amean, bmean, cmean, imean]
    sems[name] = [asem, bsem, csem, isem]
    

midpoints = np.mean(np.vstack([bins[:-1], bins[1:]]), axis=0)
midpoints

names

gnames = ['A genes (n=64)', 'B genes (n=35)', 'C genes (n=71)']

fig, axs = plt.subplots(2, 4, figsize=(5*4,4*2), sharex=True, sharey=True)

# ax.set_title('P28NR')
linestyle = '-'
for ax, name in zip(axs.flat, names):
    # (lnorm_mean, lnorm_sem, lnorm_std, n, d, db) = stats[name]
    amean, bmean, cmean, imean = means[name]
    asem, bsem, csem, isem = sems[name]
    
    x = midpoints
    ax.plot(x, amean, label='A genes', color='C0', linestyle=linestyle)
    ax.fill_between(x, amean-asem, amean+asem, color='C0', alpha=0.1, edgecolor='none')
    ax.plot(x, bmean, label='B genes', color='C1', linestyle=linestyle)
    ax.fill_between(x, bmean-bsem, bmean+bsem, color='C1', alpha=0.1, edgecolor='none')
    ax.plot(x, cmean, label='C genes', color='C2', linestyle=linestyle)
    ax.fill_between(x, cmean-csem, cmean+csem, color='C2', alpha=0.1, edgecolor='none')
    ax.axhline(color='lightgray', linestyle='dotted', zorder=1)

    sns.despine(ax=ax)
    ax.set_xticks([0, 100, 200, 300])
    ax.set_xlim(left=100, right=350)
    ax.set_ylim([-0.3, 0.3])
    ax.grid(False)
    ax.set_title(name)
axs.flat[0].set_ylabel('mean (expr. +/- sem)')

    
fig.subplots_adjust(wspace=0.1)
# powerplots.savefig_autodate(fig, outdatadir+'/grant_saumya_lineq_abc_v3.pdf')

gnames = ['IEGs']

fig, axs = plt.subplots(2, 4, figsize=(5*4,4*2), sharex=True, sharey=True)

# ax.set_title('P28NR')
linestyle = '-'
for ax, name in zip(axs.flat, names):
    (lnorm_mean, lnorm_sem, lnorm_std, n, d, db) = stats[name]
    amean, bmean, cmean, imean = means[name]
    asem, bsem, csem, isem = sems[name]
    
    x = midpoints
    ax.plot(x, imean, label='IEGs', color='C4', linestyle=linestyle)
    ax.fill_between(x, imean-isem, imean+isem, color='C0', alpha=0.1, edgecolor='none')
    ax.axhline(color='lightgray', linestyle='dotted', zorder=1)

    sns.despine(ax=ax)
    ax.set_xticks([0, 100, 200, 300])
    ax.set_xlim(left=100, right=350)
    # ax.set_ylim([-0.3, 0.3])
    ax.grid(False)
    ax.set_title(name)
axs.flat[0].set_ylabel('mean (expr. +/- sem)')

    
fig.subplots_adjust(wspace=0.1)
# powerplots.savefig_autodate(fig, outdatadir+'/grant_saumya_lineq_abc_v3.pdf')

samp_gene_dpth_mat = np.array([np.array(means[name]) for name in names]) 
print(samp_gene_dpth_mat.shape) # sample, gene group, depth

nr_mat = samp_gene_dpth_mat[:4]
nr_mean = np.mean(nr_mat, axis=0) # gene group, depth
nr_sem  = np.std(nr_mat, axis=0)/np.sqrt(4) # gene group, depth

dr_mat = samp_gene_dpth_mat[4:]
dr_mean = np.mean(dr_mat, axis=0) # gene group, depth
dr_sem  = np.std(dr_mat, axis=0)/np.sqrt(4) # gene group, depth
nr_mean.shape, dr_mean.shape

# t-test between NR and DR for each gene group and each location
from scipy import stats
from statsmodels.stats.multitest import multipletests

ts, ps = stats.ttest_ind(nr_mat, dr_mat)
rejs, qs, _, _ = multipletests(np.nan_to_num(ps, nan=1).reshape(-1,), alpha=0.05, method='fdr_bh')
qs = qs.reshape(ps.shape)
nrdr_mean = np.stack([nr_mean, dr_mean], axis=2).mean(axis=2)

# nr_bio_mat = np.stack([
#     np.mean(samp_gene_dpth_mat[0:2], axis=0),
#     np.mean(samp_gene_dpth_mat[2:4], axis=0),
#     ], axis=0)
# dr_bio_mat = np.stack([
#     np.mean(samp_gene_dpth_mat[4:6], axis=0),
#     np.mean(samp_gene_dpth_mat[6:8], axis=0),
#     ], axis=0)
# ts, ps = stats.ttest_ind(nr_bio_mat, dr_bio_mat)
# qs = ps
# rejs, qs, _, _ = multipletests(np.nan_to_num(ps, nan=1).reshape(-1,), alpha=0.05, method='fdr_bh')
# qs = qs.reshape(ps.shape)
# nrdr_mean = np.stack([nr_mean, dr_mean], axis=2).mean(axis=2)

titles = ['NR', 'DR']
data_mean = [nr_mean, dr_mean]
data_sem = [nr_sem, dr_sem]
gnames = ['A genes', 'B genes', 'C genes']
colors = ['C0', 'C1', 'C2']

fig, axs = plt.subplots(1, 2, figsize=(5*2,4), sharex=True, sharey=True)
for ax, cond_mean, cond_sem, title in zip(axs, data_mean, data_sem, titles):
    # ax.set_title('P28NR')
    linestyle = '-'
    ax.axhline(color='lightgray', linestyle='dotted', zorder=1)
    for i, (gname, color) in enumerate(zip(gnames, colors)):
        ax.plot(midpoints, cond_mean[i], label=gname, color=color, linestyle=linestyle)
        ax.fill_between(midpoints, cond_mean[i]-cond_sem[i], cond_mean[i]+cond_sem[i], color=color, alpha=0.1, edgecolor='none')

    sns.despine(ax=ax)
    ax.set_xticks([0, 100, 200, 300])
    ax.set_xlim(left=100, right=350)
    # ax.set_ylim([-0.3, 0.3])
    ax.grid(False)
    ax.set_title(title)
    ax.set_xlabel('upper->lower cortical depth')
    
axs[0].set_ylabel('mean (expr. +/- sem)')
fig.subplots_adjust(wspace=0.1)
# powerplots.savefig_autodate(fig, outdatadir+'/grant_saumya_lineq_abc_v3.pdf')
plt.show()


titles = ['NR', 'DR']
data_mean = [nr_mean, dr_mean]
data_sem = [nr_sem, dr_sem]
gnames = ['IEGs']
colors = ['C4']
i = 3

fig, axs = plt.subplots(1, 2, figsize=(5*2,4), sharex=True, sharey=True)
for ax, cond_mean, cond_sem, title in zip(axs, data_mean, data_sem, titles):
    # ax.set_title('P28NR')
    linestyle = '-'
    ax.axhline(color='lightgray', linestyle='dotted', zorder=1)
    for (gname, color) in zip(gnames, colors):
        ax.plot(midpoints, cond_mean[i], label=gname, color=color, linestyle=linestyle)
        ax.fill_between(midpoints, cond_mean[i]-cond_sem[i], cond_mean[i]+cond_sem[i], color=color, alpha=0.1, edgecolor='none')

    sns.despine(ax=ax)
    ax.set_xticks([0, 100, 200, 300])
    ax.set_xlim(left=100, right=350)
    # ax.set_ylim([-0.3, 0.3])
    ax.grid(False)
    ax.set_title(title)
    ax.set_xlabel('upper->lower cortical depth')
    
axs[0].set_ylabel('mean (expr. +/- sem)')
fig.subplots_adjust(wspace=0.1)
# powerplots.savefig_autodate(fig, outdatadir+'/grant_saumya_lineq_abc_v3.pdf')
plt.show()

linestyles = ['-', '--']
data_mean = [nr_mean, dr_mean]
data_sem = [nr_sem, dr_sem]
gnames = ['A genes', 'B genes', 'C genes']
titles = gnames
colors = ['C0', 'C1', 'C2']
labels = ['NR', 'DR']
sigs = qs
allmeans = nrdr_mean

fig, axs = plt.subplots(1, 3, figsize=(5*3,4), sharex=True, sharey=True)
for i, (ax, gname, color) in enumerate(zip(axs, gnames, colors)):
    ax.axhline(color='lightgray', linestyle='dotted', zorder=1)
    for cond_mean, cond_sem, title, linestyle in zip(data_mean, data_sem, titles, linestyles):
        ax.plot(midpoints, cond_mean[i], label=gname, color=color, linestyle=linestyle, marker='o', markersize=5)
        ax.fill_between(midpoints, cond_mean[i]-cond_sem[i], cond_mean[i]+cond_sem[i], color=color, alpha=0.1, edgecolor='none')
        
    for _x, _y, _sig in zip(midpoints, allmeans[i], sigs[i]):
        if _sig < 1e-3:
            ax.text(_x, _y, "***", ha='left', va='center', fontsize=12, rotation=90)
            ax.vlines(_x, _y-0.02, _y+0.02, color='k', linewidth=0.5)
        elif _sig < 5e-2:
            ax.text(_x, _y, "*", ha='left', va='center', fontsize=12, rotation=90)
            ax.vlines(_x, _y-0.02, _y+0.02, color='k', linewidth=0.5)

    sns.despine(ax=ax)
    ax.set_xticks([0, 100, 200, 300])
    ax.set_xlim(left=100, right=350)
    ax.set_ylim([-0.2, 0.3])
    ax.grid(False)
    ax.set_title(gname)
    ax.set_xlabel('upper->lower cortical depth')
    
axs[0].set_ylabel('mean (expr. +/- sem)')
fig.subplots_adjust(wspace=0.1)
# powerplots.savefig_autodate(fig, outdatadir+'/grant_saumya_lineq_abc_v3.pdf')
plt.show()

linestyles = ['-', '--']
data_mean = [nr_mean, dr_mean]
data_sem = [nr_sem, dr_sem]
gnames = ['IEGs']
titles = gnames
colors = ['C4']
labels = ['NR', 'DR']
i = 3
sigs = qs
allmeans = nrdr_mean

fig, ax = plt.subplots(1, 1, figsize=(5*1,4), sharex=True, sharey=True)
for (gname, color) in zip(gnames, colors):
    ax.axhline(color='lightgray', linestyle='dotted', zorder=1)
    for cond_mean, cond_sem, linestyle in zip(data_mean, data_sem, linestyles):
        ax.plot(midpoints, cond_mean[i], label=gname, color=color, linestyle=linestyle, marker='o', markersize=5)
        ax.fill_between(midpoints, cond_mean[i]-cond_sem[i], cond_mean[i]+cond_sem[i], color=color, alpha=0.1, edgecolor='none')
        
    for _x, _y, _sig in zip(midpoints, allmeans[i], sigs[i]):
        if _sig < 1e-3:
            ax.text(_x, _y, "***", ha='left', va='center', fontsize=12, rotation=90)
            ax.vlines(_x, _y-0.1, _y+0.1, color='k', linewidth=0.5)
        elif _sig < 5e-2:
            ax.text(_x, _y, "*", ha='left', va='center', fontsize=12, rotation=90)
            ax.vlines(_x, _y-0.1, _y+0.1, color='k', linewidth=0.5)

    sns.despine(ax=ax)
    ax.set_xticks([0, 100, 200, 300])
    ax.set_xlim(left=100, right=350)
    ax.set_ylim([-0.2, 1.2])
    ax.grid(False)
    ax.set_title(gname)
    ax.set_xlabel('upper->lower cortical depth')
    
ax.set_ylabel('mean (expr. +/- sem)')
fig.subplots_adjust(wspace=0.1)
# powerplots.savefig_autodate(fig, outdatadir+'/grant_saumya_lineq_abc_v3.pdf')
plt.show()

# stability analysis

def shuff_genes(mat, seed=0):
    """
    shuffle each column independently
    
    if cell by gene as input; it will shuffle each gene independently across cells
    """
    rng = np.random.default_rng(seed=seed)
    mat_shuff = rng.permuted(mat, axis=0)
    return mat_shuff

def downsample_X(X, p=0.8):
    """downsample cells (assuming it is feature by cell matrix)
    """
    return X[:,np.random.rand(X.shape[1])<p]

def aa_inference(X):
    """
    """
    XC, _, _, _, _ = PCHA(X, noc=3, delta=0)
    XC = np.array(XC)
    XC = XC[:,np.argsort(XC[0])] # assign an order according to x-axis 
    return XC

adata_nr

xi, yi, xsign, ysign = 1,2,-1,1

pcs_nr = adata_nr.obsm['pcs_typegenes'][:,[xi,yi]].dot(np.diag([xsign, ysign]))
rgbs = adata[adata_nr.obs.index].obsm['additive_rgb']
# AA inference
X = pcs_nr.T

# orig
XC = aa_inference(X) 

np.random.seed(0)
# stability - downsamp cells
nrepeats = 20
res_XC_sub = []
for i in range(nrepeats):
    X_sub = downsample_X(X, p=0.8)
    XC_sub = aa_inference(X_sub) 
    res_XC_sub.append(XC_sub)
    

mat = stats.zscore(adata_nr[:,abcgenes].layers['ljnorm'], axis=0)

np.random.seed(0)
# stability - downsamp genes
nrepeats = 20
res_XC_sub2 = []
for i in range(nrepeats):
    mat_sub = downsample_X(mat, p=0.8) # downsamp genes
    pcs = PCA(n_components=3).fit_transform(mat_sub)[:,[1,2]]
    X_sub = pcs.T
    
    xcorr = stats.pearsonr(X[0], X_sub[0])[0]
    ycorr = stats.pearsonr(X[1], X_sub[1])[0]
    if xcorr > 0:
        xsign = 1
    else:
        xsign = -1
    if ycorr > 0:
        ysign = 1
    else:
        ysign = -1
        
    X_sub = np.diag([xsign, ysign]).dot(X_sub)
    
    XC_sub = aa_inference(X_sub) 
    res_XC_sub2.append(XC_sub)

mat.shape, mat_sub.shape

# mat = stats.zscore(adata_nr[:,abcgenes].layers['ljnorm'], axis=0)
# mat_shuff = shuff_genes(mat)
# pcs_nr_shuff = PCA(n_components=5).fit_transform(mat_shuff)
# X_shuff = pcs_nr_shuff.T

# # orig
# XC_shuff, _, _, _, _ = PCHA(X_shuff, noc=3, delta=0)
# XC_shuff = np.array(XC_shuff)

# nrepeats = 20
# res_XC_sub_shuff = []
# for i in range(nrepeats):
#     X_sub = downsample_X(X_shuff, p=0.8)
#     XC_sub, _, _, _, _ = PCHA(X_sub, noc=3, delta=0)
#     XC_sub = np.array(XC_sub)
#     XC_sub = XC_sub[:, np.argsort(XC_sub[0])] # assign an order
    
#     res_XC_sub_shuff.append(XC_sub)

fig, ax = plt.subplots(1,1,figsize=(6*1,4))
ax.scatter(X[0], X[1], s=5, c=rgbs) # , cmap='coolwarm', vmin_p=5, vmax_p=95)

# add_triangle(XC, ax, vertices=True, linecolor='k', edgecolors='k', linewidths=1, marker='o', zorder=2)
for i in range(nrepeats):
    XC_i = res_XC_sub[i]
    add_triangle(XC_i, ax, zorder=0, vertices=False, linecolor='gray',)
    
ax.set_aspect('equal')
sns.despine(ax=ax)
ax.grid(False)
plt.show()

fig, ax = plt.subplots(1,1,figsize=(3*1,2))
for i in range(nrepeats):
    XC_i = res_XC_sub[i]
    add_triangle(XC_i, ax, zorder=0, vertices=True, linecolor='gray', edgecolors='k', linewidths=1, marker='o', s=30, )
    
ax.set_aspect('equal')
sns.despine(ax=ax)
ax.grid(False)
plt.show()

fig, ax = plt.subplots(1,1,figsize=(3*1,2))
for i in range(nrepeats):
    XC_i = res_XC_sub2[i]
    add_triangle(XC_i, ax, zorder=0, vertices=True, linecolor='gray', edgecolors='k', linewidths=1, marker='o', s=30, )
    
ax.set_aspect('equal')
sns.despine(ax=ax)
ax.grid(False)
plt.show()

from sklearn.metrics import pairwise_distances

XC_avg = np.mean(res_XC_sub, axis=0).T
XC_std = np.std(res_XC_sub, axis=0).T
XC_std2 = np.std(res_XC_sub2, axis=0).T
# ds = pairwise_distances(XC_avg)[np.triu_indices(3, k=1)]
ds = pairwise_distances(XC_avg)
signals = np.sort(ds, axis=1)[:,1]
epsilons = np.sqrt(np.sum(np.power(XC_std,2), axis=1))
epsilons2 = np.sqrt(np.sum(np.power(XC_std2,2), axis=1))

fig, ax = plt.subplots(figsize=(2,4))
ax.bar(0, np.mean(epsilons), color='gray', edgecolor='k')
ax.bar(1, np.mean(epsilons2), color='gray', edgecolor='k')
ax.bar(2, np.mean(ds), color='k', edgecolor='k')

ax.set_xticks([0,1,2])
# ax.set_xticklabels(['uncertainty within archetypes', 'dist. between archetypes'])
# ax.set_ylabel('distance in PC space')
ax.grid(False)
sns.despine(ax=ax)
plt.show()

np.sort(ds, axis=0)

fig, ax = plt.subplots(1,1,figsize=(6,4))
ax.scatter(X[0], X[1], s=3, c=rgbs) # , cmap='coolwarm', vmin_p=5, vmax_p=95)
add_triangle(XC, ax, vertices=True, linecolor='k', edgecolors='k', linewidths=1, marker='o', zorder=2)
# for i in range(nrepeats):
#     XC_i = res_XC_sub[i]
#     add_triangle(XC_i, ax, zorder=0, vertices=False, linecolor='gray',)
    
ax.set_aspect('equal')
sns.despine(ax=ax)
ax.grid(False)
plt.show()

gns = [agenes, bgenes, cgenes, iegs]
titles = ['A genes', 'B genes', 'C genes', 'IEGs']
adatas = [adata_nr]
conditions = ['NR']

fig, axss = plt.subplots(1,4,figsize=(4*5,3*1), sharex=True, sharey=True)
if axss.ndim==1: axss = axss.reshape(1,-1)
for i, (axs, adatasub, condition) in enumerate(zip(axss, adatas, conditions)):
    condition = conditions[i]
    for j, (ax, gn, title,) in enumerate(zip(axs, gns, titles)):
        g = adatasub[:,gn].layers['ljnorm'].mean(axis=1)
        x = xsign*adatasub.obsm['pcs_typegenes'][:,xi]
        y = ysign*adatasub.obsm['pcs_typegenes'][:,yi]
        
        p = utils_merfish.st_scatter_ax(fig, ax, x, y, gexp=g, s=5, cmap='coolwarm', vmin_p=0, vmax_p=95)
        colorbar = plt.colorbar(p, aspect=5, shrink=0.3)
        
        # Show ticks but no grid
        ax.set_aspect('equal')
        ax.axis('on')
        ax.grid(False)  # Turn off grid lines
        sns.despine(ax=ax)
        ax.tick_params(axis='both', which='both', bottom=True, left=True)
        # ax.set_xticks([-10,0,10])
        
        if i == 0:
            ax.set_title(title)
        if j == 0:
            ax.set_ylabel(condition, rotation=0, loc='top')
            
        # add the triangle
        add_triangle(XC, ax)
plt.show()
