==> archive2/analysis_merfish/47.v1_NRDR_integration.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import anndata as ad
import scanpy as sc
from scipy import stats
import os

from scipy import spatial
from scipy import sparse
from scipy.interpolate import CubicSpline
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from sklearn.neighbors import NearestNeighbors
import networkx as nx
from umap import UMAP

import json

import importlib

from scroutines import powerplots
from scroutines.miscu import is_in_polygon

import utils_merfish
importlib.reload(utils_merfish)
from utils_merfish import rot2d, st_scatter, st_scatter_ax, plot_cluster, binning
from utils_merfish import RefLineSegs

import merfish_datasets
import merfish_genesets
importlib.reload(merfish_datasets)
from merfish_datasets import merfish_datasets
from merfish_datasets import merfish_datasets_params

from scroutines import basicu

def get_qc_metrics(df):
    """
    return metrics
     - key
      - (name, val, medval, bins)
    """
    metrics = {}
    cols  = ['volume', 'gncov', 'gnnum']
    names = ['cell volume', 'num transcripts', 'num genes']
    
    for col, name in zip(cols, names):
        val = df[col].values
        medval = np.median(val)
        bins = np.linspace(0, 10*medval, 50)
        
        metrics[col] = (name, val, medval, bins)
    return metrics

def get_norm_counts(adata, scaling=500):
    """norm - equalize the volume to be 500 for all cells
    """
    cnts = adata.X
    vol = adata.obs['volume'].values
    normcnts = cnts/vol.reshape(-1,1)*scaling
    adata.layers['norm'] = normcnts
    
    return normcnts

def get_largest_spatial_components(adata, k=100, dist_th=80):
    """
    k - number of neighbors
    dist_th - distance to call connected components
    
    returns
        - indices of the largest components
    """
    XY = adata.obs[['x', 'y']].values
    nc = len(XY)

    # kNN
    nbrs = NearestNeighbors(n_neighbors=k, algorithm='auto').fit(XY)
    distances, indices = nbrs.kneighbors(XY)

    # filtered by distance th
    val = distances[:,1:].reshape(-1,)
    i = np.repeat(indices[:,0],k-1)
    j = indices[:,1:].reshape(-1,)

    indices_filtered = np.vstack([i[val < dist_th], j[val < dist_th]]).T

    G = nx.Graph()
    G.add_nodes_from(np.arange(nc))
    G.add_edges_from(indices_filtered)
    components = nx.connected_components(G)
    largest_component = max(components, key=len)
    indices_selected = np.array(list(largest_component))

    print(f"fraction of cells included: {len(largest_component)/nc: .2f}" )
    
    return indices_selected, XY

# load data and construct adata 

np.random.seed(0)

outdir     = "/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/results_merfish/plots_240718"
outdatadir = "/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish/organized"
!mkdir -p $outdir
!mkdir -p $outdatadir

genesets, df = merfish_genesets.get_all_genesets()
for key, item in genesets.items():
    print(key, len(item))

ddir = "/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish/organized" 
fout = os.path.join(ddir, 'P28NRDR_v1l23_merged_240508.h5ad')
!ls $ddir/*l2*240723.h5ad 

%%time
names = [
    'P28NRa_ant', 
    'P28NRa_pos',
    
    'P28NRb_ant', 
    'P28NRb_pos',
    
    'P28DRa_ant', 
    'P28DRa_pos',
    
    'P28DRb_ant', 
    'P28DRb_pos',
]

alldata = {}
for name in names:
    adatasub = ad.read(os.path.join(ddir, f'{name}_l2_v1_240723.h5ad')) 
    adatasub.obs.index = np.char.add(f'{name}', adatasub.obs.index.values)
    alldata[name] = adatasub 
    print(name, len(alldata[name]))
    
genes = adatasub.var.index.values
genes.shape

agenes = genesets['a']
bgenes = genesets['b']
cgenes = genesets['c']
iegs   = genesets['i']
up_agenes = genesets['a_up']
abcgenes = np.hstack([agenes, bgenes, cgenes])
len(abcgenes)

agenes_idx = basicu.get_index_from_array(adatasub.var.index.values, agenes)
bgenes_idx = basicu.get_index_from_array(adatasub.var.index.values, bgenes)
cgenes_idx = basicu.get_index_from_array(adatasub.var.index.values, cgenes)
igenes_idx = basicu.get_index_from_array(adatasub.var.index.values, iegs)

mean_total_rna_target = 250
adata_premerge = []
for i, name in enumerate(names):
    adatasub = alldata[name].copy()
    
    adatasub.obs['sample'] = name
    
    norm_cnts = adatasub.layers['norm']
    mean_per_batch = np.mean(norm_cnts.sum(axis=1))
    adatasub.layers['jnorm']  = norm_cnts*(mean_total_rna_target/mean_per_batch)
    adatasub.layers['ljnorm'] = np.log2(1+adatasub.layers['jnorm'])
    
    adatasub.obs['norm_transcript_count']  = adatasub.layers['norm'].sum(axis=1)
    adatasub.obs['jnorm_transcript_count'] = adatasub.layers['jnorm'].sum(axis=1)
    
    adata_premerge.append(adatasub)
    
adata_premerge = ad.concat(adata_premerge)

adata_merged = []
for i, name in enumerate(names):
    j = i // 4
    i = i % 4
    
    adatasub = adata_premerge[adata_premerge.obs['sample']==name].copy()
        
    adatasub.obs['sample'] = name
    adatasub.obs['depth_show'] = -adatasub.obs['depth'].values - i*1300 # name
    adatasub.obs['width_show'] =  adatasub.obs['width'].values - np.min(adatasub.obs['width'].values) + j*2500   # name
    adata_merged.append(adatasub)
    print(adatasub.shape)
    # break
    
adata_merged = ad.concat(adata_merged)


sns.histplot(adata_premerge.obs['jnorm_transcript_count'], bins=np.linspace(0, 500, 50))

sns.histplot(adata_premerge.obs['norm_transcript_count'], bins=np.linspace(0, 500, 50))

from scipy.stats import zscore

marker_genes = [
       'Ptprn', 'Slc17a7', 'Gad1', 'Fos', 
       
       'Gfap', 'Slc6a13', 'Slc47a1',
       'Grin2c', 'Aqp4', 'Rfx4', 'Sox21', 'Slc1a3',
       
       'Sox10', 'Pdgfra', 'Mog',
       
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
      ] 

genes_noniegs = np.array([g for g in genes if g not in iegs])
genes_noniegs.shape

adata = adata_merged # [:,marker_genes].copy()


# PCA
pca = PCA(n_components=50)
pcs = pca.fit_transform(zscore(adata[:,genes_noniegs].layers['ljnorm'], axis=1))
# pcs = pca.fit_transform(zscore(adata.layers['ljnorm'], axis=1))
ucs = UMAP(n_components=2, n_neighbors=30, random_state=0).fit_transform(pcs)

adata.obsm['pca'] = pcs
adata.obsm['umap'] = ucs
sc.pp.neighbors(adata, n_neighbors=30, use_rep='pca', random_state=0)

# clustering
r = 0.3
sc.tl.leiden(adata, resolution=r, key_added=f'leiden_r{r}', random_state=0, n_iterations=10)

# plot
gn = 'Slc17a7'
# gn = 'Fos'
# gn = 'Gad1'
g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))

# add some quality metrics

fig, (ax1, ax2) = plt.subplots(1,2,figsize=(2*5,1*4))
utils_merfish.st_scatter_ax(fig, ax1, pcs[:,0], pcs[:,1], gexp=g)
utils_merfish.st_scatter_ax(fig, ax2, ucs[:,0], ucs[:,1], gexp=g)
plt.show()

# plot
gn = 'Slc17a7'
# gn = 'Fos'
# gn = 'Gad1'
g = adata[:,gn].layers['jnorm'].reshape(-1,)

# add some quality metrics

fig, (ax1, ax2) = plt.subplots(1,2,figsize=(2*5,1*4))
utils_merfish.st_scatter_ax(fig, ax1, pcs[:,0], pcs[:,1], gexp=g)
utils_merfish.st_scatter_ax(fig, ax2, ucs[:,0], ucs[:,1], gexp=g)
plt.show()


clsts = adata.obs[f'leiden_r{r}'].astype(int)
xr =  adata.obs['width_show']
yr =  adata.obs['depth_show']
ux = adata.obsm['umap'][:,0]
uy = adata.obsm['umap'][:,1]
utils_merfish.plot_cluster(clsts, xr, yr, ux, uy, s=2)

samples, uniq_labels = pd.factorize(adata.obs['sample']) # .astype(int)
utils_merfish.plot_cluster(samples, xr, yr, ux, uy, s=2)

np.unique(clsts, return_counts=True)

clsts = adata.obs[f'leiden_r{r}'].astype(int)
uniq_clsts = np.unique(clsts)


for clst in uniq_clsts:
    show = (clsts == clst)
    xr =  adata.obs['width_show']
    yr =  adata.obs['depth_show']
    ux    = adata.obsm['umap'][:,0]
    uy    = adata.obsm['umap'][:,1]
    utils_merfish.plot_cluster(show, xr, yr, ux, uy, s=2, cmap=plt.cm.copper_r, suptitle=clst)

# figure out major cell population by marker genes; groups of genes; and quality metrics

# plot
marker_genes = [
       'Ptprn', 'Slc17a7', 'Gad1', 'Fos', 
       
       'Gfap', 'Slc6a13', 'Slc47a1',
       'Grin2c', 'Aqp4', 'Rfx4', 'Sox21', 'Slc1a3',
       
       'Sox10', 'Pdgfra', 'Mog',
       
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
      ] 
gns = marker_genes
n = len(gns)
nx = 4
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for gn, ax in zip(gns, axs.flat):
    g = adata[:,gn].layers['jnorm'].reshape(-1,)
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g)
    ax.set_title(gn)
plt.show()


# plot
gns = marker_genes
n = len(gns)
nx = 4
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*5,ny*4))
for gn, ax in zip(gns, axs.flat):
    g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g)
    ax.set_title(gn)
plt.show()



# RNA data basic analysis 

def preprocessing(adata):
    # filter genes
    cond = np.ravel((adata.X>0).sum(axis=0)) > 10 # expressed in more than 10 cells
    adata_sub = adata[:,cond]

    # counts
    x = adata_sub.X
    cov = adata_sub.obs['n_counts'].values

    # CP10k
    xn = (sparse.diags(1/cov).dot(x))*1e4

    # log10(CP10k+1)
    xln = xn.copy()
    xln.data = np.log10(xln.data+1)

    adata_sub.layers['norm'] = xn
    adata_sub.layers['lognorm'] = xln
    
    return adata_sub

def get_hvgs(adata, layer, nbin=20, qth=0.3):
    """
    """
    xn = adata.layers[layer]
    
    # min
    gm = np.ravel(xn.mean(axis=0))

    # var
    tmp = xn.copy()
    tmp.data = np.power(tmp.data, 2)
    gv = np.ravel(tmp.mean(axis=0))-gm**2

    # cut 
    lbl = pd.qcut(gm, nbin, labels=np.arange(nbin))
    gres = pd.DataFrame()
    gres['lbl'] = lbl
    gres['mean'] = gm
    gres['var'] = gv
    gres['ratio']= gv/gm

    # select
    gres_sel = gres.groupby('lbl')['ratio'].nlargest(int(qth*(len(gm)/nbin))) #.reset_index()
    gsel_idx = np.sort(gres_sel.index.get_level_values(1).values)

    assert np.all(gsel_idx != -1)
    
    return adata.var.index.values[gsel_idx]

#
f = '/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/cheng21_cell_scrna/organized/P28NR.h5ad'
adata_rna_raw = sc.read(f)
adata_rna_raw

adata_rna_raw.obs['Class_broad']

# adata_rna_raw = adata_rna_raw[~adata_rna_raw.obs['subclass'].isnull()]
adata_rna_raw = preprocessing(adata_rna_raw)
adata_rna_raw

hvgs = get_hvgs(adata_rna_raw, 'norm')
adata_rna = adata_rna_raw[:,hvgs]
adata_rna

# genes overlap
genes_overlap = np.intersect1d(adata_rna.var.index.values, adata.var.index.values)
genes_overlap.shape

adata_rna = adata_rna[:,genes_overlap]
adata_rna

zlognorm = stats.zscore(np.array(adata_rna.layers['lognorm'].todense()), axis=0)
pca = PCA(n_components=50, random_state=0)
pcs = pca.fit_transform(zlognorm)
adata_rna.obsm['X_pca'] = pcs

sc.pp.neighbors(adata_rna, use_rep='X_pca')
sc.tl.umap(adata_rna)

sc.pl.umap(adata_rna, color=["subclass"])

sc.pl.umap(adata_rna, color=["Class_broad"])

# integration - everything to everything - now good and takes too long

genes_overlap.shape, genes_overlap

adata_rna = adata_rna[:,genes_overlap].copy()
adata_mer = adata[adata.obs['transcript_count']>50, genes_overlap].copy()
print(len(adata_rna), len(adata), len(adata_mer))
adata_rna, adata_mer

adata_rna.obs['modality'] = 'rna'
adata_mer.obs['modality'] = 'merfish'
adata_merge = sc.concat([adata_rna, adata_mer], join='outer')
# adata_merge.obs.loc[adata_rna.obs.index, 'subclass'] = adata_rna.obs['subclass']
# adata_merge.obs.loc[adata_mer.obs.index, 'transcript_count'] = adata_mer.obs['transcript_count']

lognorm_rna = np.log10(1+np.array(adata_rna.layers['norm'].todense()))
zlognorm_rna = stats.zscore(lognorm_rna, axis=0)

lognorm_mer = np.log10(1+adata_mer.layers['norm'])
zlognorm_mer = stats.zscore(lognorm_mer, axis=0)

adata_merge.obsm['X_pca2'] = PCA(n_components=20).fit_transform(np.vstack([zlognorm_rna, zlognorm_mer])) 
adata_merge

sc.external.pp.harmony_integrate(adata_merge, 'modality', basis='X_pca2', max_iter_harmony=20)

sc.pp.neighbors(adata_merge, use_rep='X_pca_harmony')
sc.tl.umap(adata_merge) # , key='modality')
adata_merge.obs['umap1'] = adata_merge.obsm['X_umap'][:,0]
adata_merge.obs['umap2'] = adata_merge.obsm['X_umap'][:,1]

sns.scatterplot(data=adata_merge.obs.sample(frac=1, replace=False), 
                x='umap1', y='umap2', hue='modality', s=1, edgecolor='none')

# label transfer from RNA data
from sklearn.neighbors import KNeighborsClassifier

ref = adata_merge[adata_rna.obs.index].copy()
qry = adata_merge[adata_mer.obs.index].copy()
knnc = KNeighborsClassifier(n_neighbors=30)

# report distance cutoff
knnc.fit(ref.obsm['X_pca_harmony'], ref.obs['Class_broad'].values)

qry.obs['pred'] = knnc.predict(qry.obsm['X_pca_harmony'])
ref.obs['pred'] = knnc.predict(ref.obsm['X_pca_harmony'])
adata_merge.obs['pred'] = knnc.predict(adata_merge.obsm['X_pca_harmony'])

from sklearn.neighbors import NearestNeighbors

def neighbor_label_transfer(k, ref_emb, qry_emb, ref_lbl, p_cutoff=0.5, dist_cutoff=None):
    """ref vs qry neighbors
    """
    unq_lbls = np.unique(ref_lbl).astype(str) # array(['L2/3_A', 'L2/3_B', 'L2/3_C'])
    n_unq_lbls = len(unq_lbls)
    ref_n = len(ref_emb)
    qry_n = len(qry_emb)
    
    neigh = NearestNeighbors(n_neighbors=k) # , radius=0.4)
    neigh.fit(ref_emb)
    dists, idx = neigh.kneighbors(qry_emb, k, return_distance=True)
    
    raw_pred = ref_lbl[idx]

    # p
    pabc = np.empty((qry_n, n_unq_lbls))
    for i, lbl in enumerate(unq_lbls):
        p = np.sum(raw_pred==lbl, axis=1)/k
        pabc[:,i] = p

    # max
    max_pred = unq_lbls[np.argmax(pabc, axis=1)]

    # 
    gated_pred = max_pred.copy()
    cond1 = np.max( pabc, axis=1) > p_cutoff
    gated_pred[~cond1] = 'NA' 
    if dist_cutoff is not None:
        cond2 = np.max(dists, axis=1) < dist_cutoff
        gated_pred[~cond2] = 'NA' 
    
    return max_pred, gated_pred, np.max(dists, axis=1)


def neighbor_self_nonself(k, ref_emb, qry_emb):
    """ref vs qry neighbors
    """
    unq_lbls = np.unique(ref_lbl).astype(str) # array(['L2/3_A', 'L2/3_B', 'L2/3_C'])
    n_unq_lbls = len(unq_lbls)
    ref_n = len(ref_emb)
    qry_n = len(qry_emb)
    lbls = np.array([0]*ref_n+[1]*qry_n)
    
    neigh = NearestNeighbors(n_neighbors=k) # , radius=0.4)
    neigh.fit(np.vstack([ref_emb, qry_emb]))
    idx = neigh.kneighbors(qry_emb, k, return_distance=False)
    
    isself = lbls[idx]

    p = np.sum(isself, axis=1)/k

    
    return p # max_pred, gated_pred, np.max(dists, axis=1)

ref_emb = ref.obsm['X_pca_harmony']
qry_emb = qry.obsm['X_pca_harmony']
ref_lbl = ref.obs['Class_broad'].values.astype(str) # res_nr['type'].values.astype(str)
print(np.unique(ref_lbl))

k = 100
max_pred, _, dists = neighbor_label_transfer(k, ref_emb, qry_emb, ref_lbl, p_cutoff=0.5, dist_cutoff=None)
ps = neighbor_self_nonself(k, ref_emb, qry_emb)

adata_merge.obs['max_pred'] = 'NA' 
adata_merge.obs['frac_self_neighbors'] = np.nan # 'NA'
adata_merge.obs['gated_pred'] = 'NA'

adata_merge.obs.loc[qry.obs.index, 'max_pred'] = max_pred
adata_merge.obs.loc[qry.obs.index, 'frac_self_neighbors'] = ps
adata_merge.obs.loc[qry.obs.index, 'gated_pred'] = np.where(ps < 0.98, max_pred, 'NA')

adata_snrnasq = adata_merge[adata_rna.obs.index].copy() #adata_merge.obs['modality']=='merfish']
adata_merfish = adata_merge[adata_mer.obs.index].copy() #adata_merge.obs['modality']=='merfish']

adata_merfish.obs['umap1_self'] = adata_merfish.obsm['umap'][:,0]
adata_merfish.obs['umap2_self'] = adata_merfish.obsm['umap'][:,1]
adata_snrnasq, adata_merfish

sc.pl.umap(adata_merfish, color=['gated_pred'])
sc.pl.umap(adata_merfish, color=['frac_self_neighbors'], vmin=0.95)
sc.pl.umap(adata_merfish, color=['transcript_count'], vmin=0.95)

sns.color_palette('tab20')

x =   adata_merfish.obs['width']
y =  -adata_merfish.obs['depth']
ux =  adata_merfish.obsm['X_umap'][:,0]
uy =  adata_merfish.obsm['X_umap'][:,1]
clsts_lbl = adata_merfish.obs['gated_pred'].values
clsts, _ = pd.factorize(clsts_lbl) # .astype(int) # requires [0,1,2...,N]
clsts_palette, clsts_cmap = utils_merfish.generate_discrete_cmap([len(np.unique(clsts))], keys=['Set2',])

csel = 'tab20'
clsts_palette2 = {
    'Excitatory':       sns.color_palette(csel, 8)[0], # 'C0', 
    'Inhibitory':       sns.color_palette(csel, 8)[1], 
    
    'OPCs':             sns.color_palette(csel, 8)[2], 
    'Oligodendrocytes': sns.color_palette(csel, 8)[3], 
    
    'Microglia':        sns.color_palette(csel, 8)[4], 
    'Endothelial':      sns.color_palette(csel, 8)[5], 

    'Astrocytes':       sns.color_palette(csel, 8)[6], 
    'VLMCs':            sns.color_palette(csel, 8)[7],
    
    'NA': 'gray',
    
}

fig, ax = plt.subplots(1,1,figsize=(1*6,1*4))
# c = np.log10(adata_merfish.obs['transcript_count'].values)
ux = adata_merfish.obs['umap1_self'].values
uy = adata_merfish.obs['umap2_self'].values
c = adata_merfish.obs['transcript_count'].values

sorting = np.argsort(c) # [::-1]
# g = ax.scatter(ux[sorting], uy[sorting], c=c[sorting], s=1, cmap='rocket_r') 
g = ax.scatter(ux, uy, c=c<100, s=3, edgecolor='none', cmap='rocket_r') 
ax.set_aspect('equal')
ax.axis('off')
fig.colorbar(g)
plt.show()



fig, axs = plt.subplots(1,2,figsize=(8*2,6))
for ax, adata_mod, col, x, y in zip(axs, 
                              [adata_merfish, adata_snrnasq], 
                              ['gated_pred', 'Class_broad'], 
                              ['umap1', 'umap1'],
                              ['umap2', 'umap2'],
                             ):
    sns.scatterplot(data=adata_mod.obs.sample(frac=1, replace=False), 
                    x=x, y=y, 
                    hue=col, 
                    palette=clsts_palette2, hue_order=list(clsts_palette2.keys()),
                    s=5, edgecolor='none', ax=ax, rasterized=True)
    ax.axis('off')
    ax.set_aspect('equal')
    ax.legend(bbox_to_anchor=(1,1), fontsize=10)
# powerplots.savefig_autodate(fig, os.path.join(outdir, 'fig1_umap.pdf'))
fig.tight_layout()
plt.show()

fig, axs = plt.subplots(1,2,figsize=(8*2,6))
for ax, adata_mod, col, x, y in zip(axs, 
                              [adata_merfish, adata_merfish, ], 
                              ['gated_pred', 'gated_pred'], 
                              ['umap1_self', 'umap1'],
                              ['umap2_self', 'umap2'],
                             ):
    sns.scatterplot(data=adata_mod.obs.sample(frac=1, replace=False), 
                    x=x, y=y, 
                    hue=col, 
                    palette=clsts_palette2, hue_order=list(clsts_palette2.keys()),
                    s=5, edgecolor='none', ax=ax, rasterized=True)
    ax.axis('off')
    ax.set_aspect('equal')
    ax.legend(bbox_to_anchor=(1,1), fontsize=10)
# powerplots.savefig_autodate(fig, os.path.join(outdir, 'fig1_umap.pdf'))
fig.tight_layout()
plt.show()

fig, ax = plt.subplots(figsize=(8,8))
sns.scatterplot(data=adata_merfish.obs.sample(frac=1, replace=False), 
                x='width_show', y='depth_show', 
                hue='gated_pred', 
                palette=clsts_palette2, hue_order=list(clsts_palette2.keys()),
                s=2, edgecolor='none', ax=ax, rasterized=True)
ax.legend(bbox_to_anchor=(1,1), fontsize=10)
ax.axis('off')
ax.set_aspect('equal')
# ax.invert_yaxis()
# powerplots.savefig_autodate(fig, os.path.join(outdir, 'fig1_dw.pdf'))
plt.show()

dfshow = adata_merfish.obs.sample(frac=1, replace=False).copy()
types = list(clsts_palette2.keys()) # dfshow['gated_pred'].unique()
print(types)

fig, axs = plt.subplots(3,3,figsize=(8*3,8*3))
for i, ax in enumerate(axs.flat):
    thistype = types[i]
    
    ax.set_title(thistype)
    sns.scatterplot(data=dfshow, 
                    x='width_show', y='depth_show', 
                    color='lightgray',
                    s=1, edgecolor='none', ax=ax, rasterized=True)
    
    sns.scatterplot(data=dfshow[dfshow['gated_pred']==thistype], 
                    x='width_show', y='depth_show', 
                    color=clsts_palette2[thistype],
                    s=10, edgecolor='none', ax=ax, rasterized=True)
    ax.axis('off')
    ax.set_aspect('equal')
    # ax.invert_yaxis()
    
    # break
    
# ax.legend(bbox_to_anchor=(1,1), fontsize=10)
# # powerplots.savefig_autodate(fig, os.path.join(outdir, 'fig1_dw.pdf'))
# plt.show()

m =  adata_merge[adata_rna.obs.index].obs['Class_broad'].value_counts() # /len(adata_rna)
m = m/m.sum()
n =  adata_merge[adata_mer.obs.index].obs['gated_pred'].value_counts() # /len(adata_mer)).loc[m.index]
print(n.loc['NA']/n.sum())
n = n.loc[m.index]
n = n/n.sum()
# n = (n/n.sum()).loc[m.index]

m, n


fig, ax = plt.subplots()
ax.scatter(m, n, color='k')
r, p = stats.spearmanr(m.values, n.values)
for i in range(len(m)):
    ax.text(m.iloc[i], n.iloc[i], m.index.values[i], fontsize=10)
    
ax.set_title(f"Spearman r={r:.2g}")

ax.plot([0, 0.1], [0, 0.1], '--', color='gray', zorder=0)
ax.set_aspect('equal')
# ax.set_xticks([0,0.1])
# ax.set_yticks([0,0.1])
sns.despine(ax=ax)

ax.set_xlabel('subclass freq. snRNA-seq')
ax.set_ylabel('subclass freq. MERFISH')
# powerplots.savefig_autodate(fig, os.path.join(outdir, 'fig1_scatter.pdf'))
ax.set_xscale('log')
ax.set_yscale('log')
                            
plt.show()

# get astro only

adata_merged

astro_cells = adata_merfish[adata_merfish.obs['gated_pred']=='Astrocytes'].obs.index.values
adata_astro = adata_merged[astro_cells]

astro_cells.shape, adata_astro

adata = adata_astro # [:,marker_genes].copy()


# PCA
pca = PCA(n_components=5)
pcs = pca.fit_transform(zscore(adata[:,genes_noniegs].layers['ljnorm'], axis=1))
# pcs = pca.fit_transform(zscore(adata.layers['ljnorm'], axis=1))
ucs = UMAP(n_components=2, n_neighbors=30, random_state=0).fit_transform(pcs)

adata.obsm['pca'] = pcs
adata.obsm['umap'] = ucs
sc.pp.neighbors(adata, n_neighbors=30, use_rep='pca', random_state=0)

# clustering
r = 0.3
sc.tl.leiden(adata, resolution=r, key_added=f'leiden_r{r}', random_state=0, n_iterations=10)

# plot
gn = 'Slc17a7'
# gn = 'Fos'
# gn = 'Gad1'
g = np.log2(1+adata[:,gn].layers['jnorm'].reshape(-1,))

# add some quality metrics

fig, (ax1, ax2) = plt.subplots(1,2,figsize=(2*5,1*4))
utils_merfish.st_scatter_ax(fig, ax1, pcs[:,0], pcs[:,1], gexp=g)
utils_merfish.st_scatter_ax(fig, ax2, ucs[:,0], ucs[:,1], gexp=g)
plt.show()

# plot
gn = 'Slc17a7'
# gn = 'Fos'
# gn = 'Gad1'
g = adata[:,gn].layers['jnorm'].reshape(-1,)

# add some quality metrics

fig, (ax1, ax2) = plt.subplots(1,2,figsize=(2*5,1*4))
utils_merfish.st_scatter_ax(fig, ax1, pcs[:,0], pcs[:,1], gexp=g)
utils_merfish.st_scatter_ax(fig, ax2, ucs[:,0], ucs[:,1], gexp=g)
plt.show()


clsts = adata.obs[f'leiden_r{r}'].astype(int)
xr =  adata.obs['width_show']
yr =  adata.obs['depth_show']
ux = adata.obsm['umap'][:,0]
uy = adata.obsm['umap'][:,1]
utils_merfish.plot_cluster(clsts, xr, yr, ux, uy, s=2)

samples, uniq_labels = pd.factorize(adata.obs['sample']) # .astype(int)
utils_merfish.plot_cluster(samples, xr, yr, ux, uy, s=2)

adata.obs['cond'] = np.where(adata.obs['sample'].str.contains('DR'), 'DR', 'NR')
nrdr, uniq_labels = pd.factorize(adata.obs['cond']) # .astype(int)
utils_merfish.plot_cluster(nrdr, xr, yr, ux, uy, s=2)

np.unique(clsts, return_counts=True)

clsts = adata.obs[f'leiden_r{r}'].astype(int)
uniq_clsts = np.unique(clsts)

for clst in uniq_clsts:
    show = (clsts == clst)
    xr =  adata.obs['width_show']
    yr =  adata.obs['depth_show']
    ux    = adata.obsm['umap'][:,0]
    uy    = adata.obsm['umap'][:,1]
    utils_merfish.plot_cluster(show, xr, yr, ux, uy, s=2, cmap=plt.cm.copper_r, suptitle=clst)

# plot
marker_genes = [
       'Gfap', 'Aqp4', 
    
       'Slc17a7', 'Ptprn',
    
       'Grin2c', 'Rfx4', 
    
       'Sox10', 'Mog', 'Pdgfra', 'Slc6a13', 
    
       'Sox21', 'Slc1a3', 'Slc47a1',
       'Pecam1', 'Cd34' , 'Tnfrsf12a', 'Sema3c', 
       'Zfhx3', 'Pag1', 'Slco2b1', 'Cx3cr1',
    
       # 'Gad1', 'Fos', 
      ] 
gns = marker_genes
n = len(gns)
nx = 6
ny = int((n+nx-1)/nx)
# add some quality metrics
fig, axs = plt.subplots(ny,nx,figsize=(nx*3,ny*3))
for gn, ax in zip(gns, axs.flat):
    g = adata[:,gn].layers['jnorm'].reshape(-1,)
    utils_merfish.st_scatter_ax(fig, ax, ucs[:,0], ucs[:,1], gexp=g)
    ax.set_title(gn)
plt.show()

# plot
gns = ['Gfap', 'Slc6a13', 'Slc17a7', 'Grin2c', 'Aqp4', 'Rfx4']

for gn in gns:

    x  =  adata.obs['width_show']
    y  =  adata.obs['depth_show']
    g = adata[:,gn].layers['jnorm'].reshape(-1,)
    vmin = 0
    vmax = np.percentile(g, 99)
    
    adata_nr = adata[adata.obs['cond']=='NR']
    ux_nr = adata_nr.obsm['umap'][:,0]
    uy_nr = adata_nr.obsm['umap'][:,1]
    g_nr = adata_nr[:,gn].layers['jnorm'].reshape(-1,)
    
    adata_dr = adata[adata.obs['cond']=='DR']
    ux_dr = adata_dr.obsm['umap'][:,0]
    uy_dr = adata_dr.obsm['umap'][:,1]
    g_dr = adata_dr[:,gn].layers['jnorm'].reshape(-1,)

    fig, (ax1, ax2, ax3) = plt.subplots(1,3,figsize=(3*3,1*3))
    utils_merfish.st_scatter_ax(fig, ax1, x, y, gexp=g, vmin=vmin, vmax=vmax)
    utils_merfish.st_scatter_ax(fig, ax2, ux_nr, uy_nr, gexp=g_nr, vmin=vmin, vmax=vmax)
    utils_merfish.st_scatter_ax(fig, ax3, ux_dr, uy_dr, gexp=g_dr, vmin=vmin, vmax=vmax)
    ax2.set_title(gn)
    plt.show()
    

sns.boxplot([g_nr, g_dr])

gn = 'Rfx4'
glist = []
for sample in names:
    g = adata[adata.obs['sample']==sample, gn].layers['jnorm'].reshape(-1,)
    glist.append(g)
sns.boxplot(glist)

def test(a, b, ax, m=0.05):
    """
    """
    s, p = stats.mannwhitneyu(a, b)
    if p > 0.05:
        mark = 'ns'
    elif p < 0.05 and p > 0.001:
        mark = '*'
    elif p < 0.001:
        mark = '***'

    # statistical annotation
    x1, x2 = 0, 1   
    y = np.max([a, b])
    y, h = y+y*m, y*m
    
    ax.plot([x1, x1, x2, x2], [y, y+h, y+h, y], lw=1.5, c='k')
    ax.text((x1+x2)*.5, y+h, mark, ha='center', va='bottom', color='k')
    
    return s, p

def check_gene(adata, gn, ax):
    glist = []
    for sample in names:
        g = np.array(adata[adata.obs['sample']==sample, gn].layers['jnorm'].reshape(-1,))
        gm = np.mean(g)
        glist.append(gm)
    gnr = glist[:4]
    gdr = glist[4:]

    ax.bar([0,1], [np.mean(gnr), np.mean(gdr)], facecolor='none', edgecolor=['C1', 'k',], linewidth=3)
    sns.swarmplot([gnr, gdr], palette=['C1', 'k'], ax=ax)
    s, p = test(gnr, gdr, ax)
    sns.despine(ax=ax)
    ax.set_title(gn)
    ax.set_xticks([0,1])
    ax.set_xticklabels(['NR', 'DR'])
    ax.set_ylabel('norm expr')

adata_rfx4 = adata[adata.obs['leiden_r0.3']=='3']
adata_rfx4

fig, ax = plt.subplots(figsize=(2,4))
check_gene(adata, 'Rfx4', ax)
plt.show()

fig, ax = plt.subplots(figsize=(2,4))
check_gene(adata_rfx4, 'Rfx4', ax)
plt.show()

# rfx4 -> target genes
rfx4_target_genes = [
    "Rfx4",
    "Junb",
    "Fosl2",
    "Mertk",
    "Rora",
    "Sdc4",
    "Nr1d1",
    "Slc1a3",
    "Nrxn1",
    "Sox21",
    "Id3",
    "Stat2",
    "Klf3",
]

n = len(rfx4_target_genes) 

fig, axs = plt.subplots(1, n, figsize=(2*n,4))
for gn, ax in zip(rfx4_target_genes, axs):
    check_gene(adata, gn, ax)
    ax.set_ylabel('')
    
axs[0].set_ylabel('norm expr')
fig.tight_layout()
plt.show()

n = len(rfx4_target_genes) 

fig, axs = plt.subplots(1, n, figsize=(2*n,4))
for gn, ax in zip(rfx4_target_genes, axs):
    check_gene(adata_rfx4, gn, ax)
    ax.set_ylabel('')
    
axs[0].set_ylabel('norm expr')
fig.tight_layout()
plt.show()

# plot
gns = rfx4_target_genes # ['Gfap', 'Slc6a13', 'Slc17a7', 'Grin2c', 'Aqp4', 'Rfx4']

for gn in gns:

    x  =  adata.obs['width_show']
    y  =  adata.obs['depth_show']
    g = adata[:,gn].layers['jnorm'].reshape(-1,)
    vmin = 0
    vmax = np.percentile(g, 98)
    
    adata_nr = adata[adata.obs['cond']=='NR']
    ux_nr = adata_nr.obsm['umap'][:,0]
    uy_nr = adata_nr.obsm['umap'][:,1]
    g_nr = adata_nr[:,gn].layers['jnorm'].reshape(-1,)
    
    adata_dr = adata[adata.obs['cond']=='DR']
    ux_dr = adata_dr.obsm['umap'][:,0]
    uy_dr = adata_dr.obsm['umap'][:,1]
    g_dr = adata_dr[:,gn].layers['jnorm'].reshape(-1,)

    fig, (ax1, ax2, ax3) = plt.subplots(1,3,figsize=(3*3,1*3))
    utils_merfish.st_scatter_ax(fig, ax1, x, y, gexp=g, vmin=vmin, vmax=vmax, s=3)
    utils_merfish.st_scatter_ax(fig, ax2, ux_nr, uy_nr, gexp=g_nr, vmin=vmin, vmax=vmax, s=3)
    utils_merfish.st_scatter_ax(fig, ax3, ux_dr, uy_dr, gexp=g_dr, vmin=vmin, vmax=vmax, s=3)
    ax2.set_title(gn)
    plt.show()
    

# add single cell level
