==> archive2/analysis_merfish/65.organize_data_presubmission.ipynb <==
from merfish_datasets import merfish_datasets
import os
import shutil

merfish_datasets

samples = [
    'P28NRa_ant',
    'P28NRa_pos',
    'P28NRb_ant',
    'P28NRb_pos',
    
    'P28DRa_ant',
    'P28DRa_pos',
    'P28DRb_ant',
    'P28DRb_pos',
]

src_dir = '/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish/' 
dst_dir = '/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish/nrdr_merfish_paper_2024/'

# for sample in samples:
#     path = merfish_datasets[sample]
#     print(sample, path)
#     src = os.path.join(src_dir,   path, 'cell_metadata.csv')
#     dst = os.path.join(dst_dir, sample, 'cell_metadata.csv')
#     shutil.copyfile(src, dst)

# for sample in samples:
#     path = merfish_datasets[sample]
#     print(sample, path)
#     src = os.path.join(src_dir,   path, 'cell_by_gene.csv')
#     dst = os.path.join(dst_dir, sample, 'cell_by_gene.csv')
#     shutil.copyfile(src, dst)
    
for sample in samples:
    path = merfish_datasets[sample]
    print(sample, path)
    src = os.path.join(src_dir,   path, 'detected_transcripts.csv')
    dst = os.path.join(dst_dir, sample, 'detected_transcripts.csv')
    shutil.copyfile(src, dst)
